refactor(coordinator/registry): Phase 3.1 — split registry.go & scheduler.go into per-domain files
Pure-move split (verified deletion-only on the sources; no body/comment/signature
changed; mutex discipline preserved by construction):
  registry.go   2809 -> 93 lines  (+ pending_request, provider, concurrency,
    privacy_policy, trust, catalog, clamp, persistence, lifecycle, trust_state,
    model_load, counters, snapshots, scoring)
  scheduler.go   947 -> 3 lines   (+ routing_constants, routing_types, reserve,
    routing_log, snapshot, admission, cost, tps, quick_capacity, drain)

scoring.go isolates the legacy FindProvider/FindProviderWithTrust/ScoreProvider
trio (zero production callers) ahead of their removal in 3.4.

Build, vet, gofmt, go test ./coordinator/registry/... -race all green.
