+++ coordinator/registry/clamp.go
@@ -1,0 +1,116 @@
+package registry
+
+import (
+	"log/slog"
+	"math"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+)
+
+// Sanity caps on provider-reported stats. A malicious (or broken) provider
+// could otherwise report absurd values to monopolize routing. These caps are
+// ~3-4x current hardware ceilings (M2 Ultra is ~800 GB/s, MLX decode is ~120
+// tok/s, max Mac Studio RAM is 512 GB) so legitimate future hardware isn't
+// clamped unnecessarily.
+const (
+	maxDecodeTPS                    = 500.0
+	maxPrefillTPS                   = 5000.0
+	maxMemoryBandwidthGBs           = 2000.0
+	maxMemoryGB                     = 1024
+	maxMemoryGBFloat                = 1024.0
+	maxReportedMaxConcurrency       = 24
+	maxTokensPotential              = 1_000_000
+	maxTokenBudgetCap         int64 = 10_000_000_000 // 10 billion — generous safety valve for total token budget capacity
+)
+
+// clampNonNeg returns v clamped into [0, max]; NaN/negative become 0.
+// The bool is true if the value was out of range.
+func clampNonNeg(v, max float64) (float64, bool) {
+	if math.IsNaN(v) || v < 0 {
+		return 0, true
+	}
+	if v > max {
+		return max, true
+	}
+	return v, false
+}
+
+// clampBackendCapacity applies sanity caps to provider-reported backend
+// capacity fields that feed the routing scorer. A provider reporting
+// TotalMemoryGB=1e9 would make gpuUtil ~= 0 and dodge health penalties, so
+// we cap it at maxMemoryGBFloat. Same for MaxTokensPotential which directly
+// controls backlog cost. NaN/negative become 0.
+func clampBackendCapacity(logger *slog.Logger, providerID string, bc *protocol.BackendCapacity) {
+	if bc == nil {
+		return
+	}
+	if v, changed := clampNonNeg(bc.TotalMemoryGB, maxMemoryGBFloat); changed {
+		logger.Warn("provider total_memory_gb out of range, clamping",
+			"provider_id", providerID, "reported", bc.TotalMemoryGB, "clamped", v)
+		bc.TotalMemoryGB = v
+	}
+	if v, changed := clampNonNeg(bc.GPUMemoryActiveGB, maxMemoryGBFloat); changed {
+		logger.Warn("provider gpu_memory_active_gb out of range, clamping",
+			"provider_id", providerID, "reported", bc.GPUMemoryActiveGB, "clamped", v)
+		bc.GPUMemoryActiveGB = v
+	}
+	if v, changed := clampNonNeg(bc.GPUMemoryPeakGB, maxMemoryGBFloat); changed {
+		bc.GPUMemoryPeakGB = v
+	}
+	if v, changed := clampNonNeg(bc.GPUMemoryCacheGB, maxMemoryGBFloat); changed {
+		bc.GPUMemoryCacheGB = v
+	}
+	for i := range bc.Slots {
+		s := &bc.Slots[i]
+		if s.MaxTokensPotential < 0 || s.MaxTokensPotential > maxTokensPotential {
+			logger.Warn("provider slot max_tokens_potential out of range, clamping",
+				"provider_id", providerID, "model", s.Model, "reported", s.MaxTokensPotential)
+			if s.MaxTokensPotential < 0 {
+				s.MaxTokensPotential = 0
+			} else {
+				s.MaxTokensPotential = maxTokensPotential
+			}
+		}
+		if s.NumRunning < 0 {
+			s.NumRunning = 0
+		}
+		if s.NumWaiting < 0 {
+			s.NumWaiting = 0
+		}
+		if s.MaxConcurrency < 0 || s.MaxConcurrency > maxReportedMaxConcurrency {
+			logger.Warn("provider slot max_concurrency out of range, clamping",
+				"provider_id", providerID, "model", s.Model, "reported", s.MaxConcurrency)
+			if s.MaxConcurrency < 0 {
+				s.MaxConcurrency = 0
+			} else {
+				s.MaxConcurrency = maxReportedMaxConcurrency
+			}
+		}
+		if v, changed := clampNonNeg(s.ObservedDecodeTPS, maxDecodeTPS); changed {
+			logger.Warn("provider slot observed_decode_tps out of range, clamping",
+				"provider_id", providerID, "model", s.Model, "reported", s.ObservedDecodeTPS, "clamped", v)
+			s.ObservedDecodeTPS = v
+		}
+		if s.ActiveTokenBudgetUsed < 0 || s.ActiveTokenBudgetUsed > maxTokenBudgetCap {
+			if s.ActiveTokenBudgetUsed < 0 {
+				s.ActiveTokenBudgetUsed = 0
+			} else {
+				s.ActiveTokenBudgetUsed = maxTokenBudgetCap
+			}
+		}
+		if s.ActiveTokenBudgetMax < 0 || s.ActiveTokenBudgetMax > maxTokenBudgetCap {
+			if s.ActiveTokenBudgetMax < 0 {
+				s.ActiveTokenBudgetMax = 0
+			} else {
+				s.ActiveTokenBudgetMax = maxTokenBudgetCap
+			}
+		}
+		if s.QueuedTokenBudget < 0 || s.QueuedTokenBudget > maxTokenBudgetCap {
+			if s.QueuedTokenBudget < 0 {
+				s.QueuedTokenBudget = 0
+			} else {
+				s.QueuedTokenBudget = maxTokenBudgetCap
+			}
+		}
+	}
+}
