+++ coordinator/registry/cost.go
@@ -1,0 +1,135 @@
+package registry
+
+import (
+	"math"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+)
+
+// buildCandidateWithReason returns the candidate plus, on rejection,
+// the reason so callers can split metrics by failure mode.
+func (r *Registry) buildCandidateWithReason(snap routingSnapshot, pr *PendingRequest) (*routingCandidate, candidateRejection, bool) {
+	statePenalty, eligible := slotStatePenalty(snap.slotState)
+	if !eligible {
+		return nil, rejectNone, false
+	}
+	if !snap.hasHeadroom {
+		return nil, rejectCapacity, false
+	}
+
+	if snap.systemMetrics.ThermalState == "critical" {
+		return nil, rejectNone, false
+	}
+
+	reqMax := pr.RequestedMaxTokens
+	if reqMax <= 0 {
+		reqMax = defaultRequestedMaxTokens
+	}
+	reqPrompt := pr.EstimatedPromptTokens
+	if reqPrompt < 0 {
+		reqPrompt = 0
+	}
+
+	// Free-memory admission gate (Phase 1). A provider that claims to
+	// serve the model but doesn't have headroom for weights + KV cache
+	// is rejected here so we don't OOM the backend post-routing.
+	if !freeMemoryAdmits(snap, reqPrompt, reqMax) {
+		return nil, rejectCapacity, false
+	}
+
+	effectiveQueue := snap.pendingForModel
+	backendDepth := snap.backendRunning + snap.backendWaiting
+	if backendDepth > effectiveQueue {
+		effectiveQueue = backendDepth
+	}
+
+	waitingBacklogTokens := float64(snap.backendWaiting * reqMax)
+	unaccountedPendingTokens := float64(snap.pendingMaxTokens) - float64(snap.maxTokensPotential) - waitingBacklogTokens
+	if unaccountedPendingTokens < 0 {
+		unaccountedPendingTokens = 0
+	}
+
+	effectiveTPS := resolveEffectiveTPS(snap)
+
+	queueMs := float64(effectiveQueue) * queueDepthPenaltyMs
+	pendingMs := float64(snap.totalPending) * totalPendingPenaltyMs
+	var backlogMs float64
+	if snap.activeTokenBudgetMax > 0 {
+		tokensAhead := float64(snap.activeTokenBudgetUsed) + float64(snap.queuedTokenBudget)
+		backlogMs = tokensAhead / effectiveTPS * 1000.0
+	} else {
+		backlogMs = backlogTokenMs(snap.maxTokensPotential, waitingBacklogTokens, unaccountedPendingTokens, effectiveTPS)
+	}
+	thisReqMs := float64(reqPrompt)/snap.prefillTPS*1000.0 + float64(reqMax)/effectiveTPS*1000.0
+	healthMs := healthPenaltyMs(snap.systemMetrics, snap.gpuMemoryActiveGB, snap.totalMemoryGB)
+	cost := statePenalty + queueMs + pendingMs + backlogMs + thisReqMs + healthMs
+
+	return &routingCandidate{
+		provider:       snap.provider,
+		snapshot:       snap,
+		costMs:         cost,
+		effectiveQueue: effectiveQueue,
+		effectiveTPS:   effectiveTPS,
+		breakdown: costBreakdown{
+			StateMs:   statePenalty,
+			QueueMs:   queueMs,
+			PendingMs: pendingMs,
+			BacklogMs: backlogMs,
+			ThisReqMs: thisReqMs,
+			HealthMs:  healthMs,
+			Total:     cost,
+		},
+	}, rejectNone, true
+}
+
+func slotStatePenalty(state string) (float64, bool) {
+	switch state {
+	case "", "running", "idle":
+		return slotStatePenaltyRunning, true
+	case "unknown":
+		// Model is available but not loaded. The provider must evict the
+		// current model and load this one — typically 15–60 seconds for
+		// large models (depends on model size and disk speed). Warm
+		// providers are strongly preferred but cold providers are still
+		// eligible when no warm alternative exists.
+		return slotStatePenaltyUnknown, true
+	case "idle_shutdown":
+		return slotStatePenaltyIdleShutdown, true
+	case "reloading", "crashed":
+		return math.Inf(1), false
+	default:
+		return slotStatePenaltyUnknown, true
+	}
+}
+
+func backlogTokenMs(maxTokensPotential int64, waitingTokens, unaccountedPendingTokens, decodeTPS float64) float64 {
+	if decodeTPS <= 0 {
+		decodeTPS = 1.0
+	}
+	totalTokensAhead := float64(maxTokensPotential) + waitingTokens + unaccountedPendingTokens
+	if totalTokensAhead < 0 {
+		totalTokensAhead = 0
+	}
+	return totalTokensAhead / decodeTPS * 1000.0
+}
+
+func healthPenaltyMs(m protocol.SystemMetrics, gpuActiveGB, totalMemGB float64) float64 {
+	penalty := m.MemoryPressure*memoryPressurePenaltyMs + m.CPUUsage*cpuUsagePenaltyMs
+	switch m.ThermalState {
+	case "fair":
+		penalty += thermalPenaltyFairMs
+	case "serious":
+		penalty += thermalPenaltySeriousMs
+	}
+	if totalMemGB > 0 {
+		gpuUtil := gpuActiveGB / totalMemGB
+		if gpuUtil < 0 {
+			gpuUtil = 0
+		}
+		if gpuUtil > 1 {
+			gpuUtil = 1
+		}
+		penalty += gpuUtil * gpuUtilizationPenaltyMs
+	}
+	return penalty
+}
