+++ coordinator/registry/counters.go
@@ -1,0 +1,99 @@
+package registry
+
+import (
+	"sync/atomic"
+)
+
+// ProviderCount returns the number of registered providers.
+// modelProviderInc increments the provider count for a model. Must be called
+// with r.mu held.
+func (r *Registry) modelProviderInc(model string) {
+	r.modelProvidersMu.Lock()
+	c, ok := r.modelProviders[model]
+	if !ok {
+		c = &atomic.Int64{}
+		r.modelProviders[model] = c
+	}
+	r.modelProvidersMu.Unlock()
+	c.Add(1)
+}
+
+// modelProviderDec decrements the provider count for a model. Must be called
+// with r.mu held.
+func (r *Registry) modelProviderDec(model string) {
+	r.modelProvidersMu.Lock()
+	c, ok := r.modelProviders[model]
+	r.modelProvidersMu.Unlock()
+	if ok {
+		v := c.Add(-1)
+		if v <= 0 {
+			r.modelProvidersMu.Lock()
+			delete(r.modelProviders, model)
+			r.modelProvidersMu.Unlock()
+		}
+	}
+}
+
+// OnlineCount returns the number of online providers.
+func (r *Registry) OnlineCount() int64 {
+	return r.onlineCount.Load()
+}
+
+// ModelProviderSnapshot returns a snapshot of model_id -> provider count.
+func (r *Registry) ModelProviderSnapshot() map[string]int64 {
+	r.modelProvidersMu.Lock()
+	snap := make(map[string]int64, len(r.modelProviders))
+	for model, c := range r.modelProviders {
+		if v := c.Load(); v > 0 {
+			snap[model] = v
+		}
+	}
+	r.modelProvidersMu.Unlock()
+	return snap
+}
+
+func (r *Registry) ProviderCount() int {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	return len(r.providers)
+}
+
+func (r *Registry) ProviderCountByVersion() map[string]int {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	counts := make(map[string]int)
+	for _, p := range r.providers {
+		p.mu.Lock()
+		online := p.Status != StatusOffline && p.Status != StatusUntrusted
+		p.mu.Unlock()
+		if !online {
+			continue
+		}
+		ver := p.Version
+		if ver == "" {
+			ver = "unknown"
+		}
+		counts[ver]++
+	}
+	return counts
+}
+
+// ForEachProvider iterates over all registered providers (read lock held).
+func (r *Registry) ForEachProvider(fn func(p *Provider)) {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	for _, p := range r.providers {
+		fn(p)
+	}
+}
+
+// ProviderIDs returns the IDs of all registered providers.
+func (r *Registry) ProviderIDs() []string {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	ids := make([]string, 0, len(r.providers))
+	for id := range r.providers {
+		ids = append(ids, id)
+	}
+	return ids
+}
