+++ coordinator/registry/lifecycle.go
@@ -1,0 +1,357 @@
+package registry
+
+import (
+	"context"
+	"encoding/base64"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/saferun"
+	"nhooyr.io/websocket"
+)
+
+// Register adds a new provider to the registry, returning its assigned ID.
+// Provider-reported model inventory is preserved even when the current catalog
+// denies every model; catalog checks are applied dynamically during routing so
+// providers that connect before a model is promoted become routable immediately
+// after the catalog is updated.
+func (r *Registry) Register(id string, conn *websocket.Conn, msg *protocol.RegisterMessage) *Provider {
+	// Clamp provider-reported performance stats used in routing score.
+	// Refuse to trust unbounded values — a malicious provider reporting
+	// DecodeTPS=1e9 would otherwise starve all other providers.
+	if v, changed := clampNonNeg(msg.DecodeTPS, maxDecodeTPS); changed {
+		r.logger.Warn("provider decode_tps out of range, clamping",
+			"provider_id", id, "reported", msg.DecodeTPS, "clamped", v)
+		msg.DecodeTPS = v
+	}
+	if v, changed := clampNonNeg(msg.PrefillTPS, maxPrefillTPS); changed {
+		r.logger.Warn("provider prefill_tps out of range, clamping",
+			"provider_id", id, "reported", msg.PrefillTPS, "clamped", v)
+		msg.PrefillTPS = v
+	}
+	if v, changed := clampNonNeg(msg.Hardware.MemoryBandwidthGBs, maxMemoryBandwidthGBs); changed {
+		r.logger.Warn("provider memory_bandwidth_gbs out of range, clamping",
+			"provider_id", id, "reported", msg.Hardware.MemoryBandwidthGBs, "clamped", v)
+		msg.Hardware.MemoryBandwidthGBs = v
+	}
+	if msg.Hardware.MemoryGB < 0 || msg.Hardware.MemoryGB > maxMemoryGB {
+		r.logger.Warn("provider memory_gb out of range, clamping",
+			"provider_id", id, "reported", msg.Hardware.MemoryGB)
+		if msg.Hardware.MemoryGB < 0 {
+			msg.Hardware.MemoryGB = 0
+		} else {
+			msg.Hardware.MemoryGB = maxMemoryGB
+		}
+	}
+
+	models := msg.Models
+
+	// Validate X25519 public key if provided.
+	// Reject invalid keys at registration rather than failing at encryption time.
+	pubKey := msg.PublicKey
+	if pubKey != "" {
+		decoded, err := base64.StdEncoding.DecodeString(pubKey)
+		if err != nil || len(decoded) != 32 {
+			r.logger.Warn("provider public key invalid, clearing",
+				"provider_id", id,
+				"error", "must be 32-byte base64-encoded X25519 key",
+			)
+			pubKey = "" // clear so provider can register but won't receive encrypted requests
+		}
+	}
+
+	p := &Provider{
+		ID:                      id,
+		Hardware:                msg.Hardware,
+		Models:                  models,
+		Backend:                 msg.Backend,
+		PublicKey:               pubKey,
+		EncryptedResponseChunks: msg.EncryptedResponseChunks,
+		PrefillTPS:              msg.PrefillTPS,
+		DecodeTPS:               msg.DecodeTPS,
+		TrustLevel:              TrustNone,
+		RuntimeVerified:         true,  // default to verified; API layer sets false when manifest check fails
+		RuntimeManifestChecked:  true,  // default to true; API layer sets false when no manifest is configured
+		ChallengeVerifiedSIP:    false, // starts false; set true by attestation challenge handler after SIP check
+		PrivacyCapabilities:     msg.PrivacyCapabilities,
+		TemplateHashes:          CloneStringMap(msg.TemplateHashes),
+		Status:                  StatusOnline,
+		Conn:                    conn,
+		LastHeartbeat:           time.Now(),
+		Reputation:              NewReputation(),
+		pendingReqs:             make(map[string]*PendingRequest),
+	}
+
+	r.mu.Lock()
+	r.providers[id] = p
+	r.onlineCount.Add(1)
+	for _, m := range models {
+		r.modelProviderInc(m.ID)
+	}
+	r.mu.Unlock()
+
+	r.logger.Info("provider registered",
+		"provider_id", id,
+		"chip", msg.Hardware.ChipName,
+		"memory_gb", msg.Hardware.MemoryGB,
+		"models", len(msg.Models),
+		"backend", msg.Backend,
+		"prefill_tps", msg.PrefillTPS,
+		"decode_tps", msg.DecodeTPS,
+	)
+
+	// Persist provider record to store (async).
+	r.persistProviderNow(p)
+
+	return p
+}
+
+func CloneStringMap(in map[string]string) map[string]string {
+	if len(in) == 0 {
+		return nil
+	}
+	out := make(map[string]string, len(in))
+	for k, v := range in {
+		out[k] = v
+	}
+	return out
+}
+
+// DisconnectDuplicatesBySerial disconnects all providers that share the same
+// serial number as the given provider, except the given provider itself.
+// This prevents multiple WebSocket connections from the same physical machine
+// from competing for the same vllm-mlx backend on localhost.
+func (r *Registry) DisconnectDuplicatesBySerial(keepID string, serial string) {
+	if serial == "" {
+		return
+	}
+
+	var toEvict []string
+
+	r.mu.RLock()
+	for id, p := range r.providers {
+		if id == keepID {
+			continue
+		}
+		if p.AttestationResult != nil && p.AttestationResult.SerialNumber == serial {
+			toEvict = append(toEvict, id)
+		}
+	}
+	r.mu.RUnlock()
+
+	for _, id := range toEvict {
+		r.mu.RLock()
+		p := r.providers[id]
+		r.mu.RUnlock()
+
+		r.logger.Warn("evicting duplicate provider from same device",
+			"evicted_id", id,
+			"kept_id", keepID,
+			"serial", serial,
+		)
+		r.Disconnect(id)
+
+		if p != nil && p.Conn != nil {
+			p.Conn.Close(websocket.StatusNormalClosure, "replaced by new connection from same device")
+		}
+	}
+}
+
+// Heartbeat updates the provider's status and stats.
+func (r *Registry) Heartbeat(id string, msg *protocol.HeartbeatMessage) {
+	r.mu.RLock()
+	p, ok := r.providers[id]
+	r.mu.RUnlock()
+	if !ok {
+		r.logger.Warn("heartbeat from unknown provider", "provider_id", id)
+		return
+	}
+
+	// Clamp heartbeat-reported capacity and metrics so a malicious provider
+	// can't skew routing by reporting absurd values (e.g. TotalMemoryGB=1e9
+	// would drive gpuUtil to 0 and sidestep health penalties).
+	clampBackendCapacity(r.logger, id, msg.BackendCapacity)
+	if v, changed := clampNonNeg(msg.SystemMetrics.MemoryPressure, 1.0); changed {
+		msg.SystemMetrics.MemoryPressure = v
+	}
+	if v, changed := clampNonNeg(msg.SystemMetrics.CPUUsage, 1.0); changed {
+		msg.SystemMetrics.CPUUsage = v
+	}
+
+	p.mu.Lock()
+	p.LastHeartbeat = time.Now()
+	p.Stats.RequestsServed += cumulativeDelta(p.lastSessionStats.RequestsServed, msg.Stats.RequestsServed)
+	p.Stats.TokensGenerated += cumulativeDelta(p.lastSessionStats.TokensGenerated, msg.Stats.TokensGenerated)
+	p.lastSessionStats = msg.Stats
+	p.SystemMetrics = msg.SystemMetrics
+	// Update backend capacity from heartbeat. A nil report clears prior live
+	// capacity so stale slot state cannot keep influencing routing.
+	p.BackendCapacity = msg.BackendCapacity
+	if p.BackendCapacity != nil {
+		chipFamily := p.Hardware.ChipFamily
+		for _, slot := range p.BackendCapacity.Slots {
+			if slot.ObservedDecodeTPS > 0 {
+				r.tpsRegistry.Record(slot.Model, chipFamily, slot.ObservedDecodeTPS)
+			}
+		}
+	}
+	// Update warm models from heartbeat. Always overwrite -- an empty list
+	// means the provider has no models loaded, and stale entries must be
+	// cleared to prevent TriggerModelSwaps from suppressing needed swaps.
+	p.WarmModels = msg.WarmModels
+	if msg.ActiveModel != nil {
+		p.CurrentModel = *msg.ActiveModel
+	} else {
+		// nil active_model means no model is loaded — clear stale state
+		// so attestation challenges don't compare against an unloaded model.
+		p.CurrentModel = ""
+	}
+	// Only update status from heartbeat if provider is not actively serving
+	// (serving status is managed by request lifecycle). Crucially, an
+	// untrusted provider must NOT transition back to StatusOnline here —
+	// that would cause an onlineCount double-decrement when Disconnect
+	// later sees StatusOnline and decrements a second time.
+	if p.Status == StatusUntrusted {
+		// no status transitions allowed
+	} else if p.Status != StatusServing || msg.Status == "idle" {
+		switch msg.Status {
+		case "idle":
+			p.Status = StatusOnline
+		case "serving":
+			p.Status = StatusServing
+		}
+	}
+	p.mu.Unlock()
+
+	r.PersistProviderThrottled(p)
+
+	// Heartbeats can make a recovered slot routable again (for example after a
+	// crash auto-restart). Drain matching queues using the canonical scheduler
+	// rather than the legacy direct queue assignment path.
+	r.drainQueuedRequestsForModels(providerModelIDs(p))
+
+	// If queue drain didn't satisfy all pending requests (no warm provider),
+	// check if a cold provider should swap models to serve queued demand.
+	r.TriggerModelSwaps()
+}
+
+func cumulativeDelta(previous, current int64) int64 {
+	if current <= 0 {
+		return 0
+	}
+	if current >= previous {
+		return current - previous
+	}
+	// The provider process restarted and reset its in-memory counters.
+	return current
+}
+
+// Disconnect removes a provider from the registry and cleans up pending requests.
+func (r *Registry) Disconnect(id string) {
+	r.mu.Lock()
+	p, ok := r.providers[id]
+	if ok {
+		delete(r.providers, id)
+		// Clear any pending model load entries for this provider.
+		for key := range r.pendingModelLoads {
+			if len(key) > len(id)+1 && key[:len(id)+1] == id+":" {
+				delete(r.pendingModelLoads, key)
+			}
+		}
+		p.mu.Lock()
+		if p.Status != StatusUntrusted {
+			r.onlineCount.Add(-1)
+			for _, m := range p.Models {
+				r.modelProviderDec(m.ID)
+			}
+		}
+		p.mu.Unlock()
+	}
+	r.mu.Unlock()
+
+	if !ok {
+		return
+	}
+
+	// Close all pending request channels so consumers get errors.
+	p.mu.Lock()
+	for reqID, pr := range p.pendingReqs {
+		pr.ErrorCh <- protocol.InferenceErrorMessage{
+			Type:       protocol.TypeInferenceError,
+			RequestID:  reqID,
+			Error:      "provider disconnected",
+			StatusCode: 502,
+		}
+		close(pr.ChunkCh)
+		close(pr.CompleteCh)
+		close(pr.ErrorCh)
+	}
+	p.pendingReqs = make(map[string]*PendingRequest)
+	p.mu.Unlock()
+
+	r.logger.Info("provider disconnected", "provider_id", id)
+}
+
+// GetProvider returns a provider by ID, or nil if not found.
+func (r *Registry) GetProvider(id string) *Provider {
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+	return r.providers[id]
+}
+
+// SetProviderIdle updates a provider's status after a request completes.
+// If pending count reaches zero, status goes back to online. If there are
+// queued requests and the provider has concurrency headroom, the next
+// queued request is assigned immediately.
+func (r *Registry) SetProviderIdle(id string) {
+	r.mu.RLock()
+	p, ok := r.providers[id]
+	r.mu.RUnlock()
+	if !ok {
+		return
+	}
+
+	p.mu.Lock()
+	if p.pendingCount() == 0 && p.Status != StatusUntrusted && p.Status != StatusOffline {
+		p.Status = StatusOnline
+	}
+	p.mu.Unlock()
+
+	// Use all newly available capacity, not just a single queued request.
+	r.drainQueuedRequestsForModels(providerModelIDs(p))
+}
+
+// StartEvictionLoop starts a background goroutine that removes providers
+// that haven't sent a heartbeat within the given timeout. It stops when
+// the context is cancelled.
+func (r *Registry) StartEvictionLoop(ctx context.Context, timeout time.Duration) {
+	ticker := time.NewTicker(timeout / 3)
+	saferun.Go(r.logger, "registry.evictionLoop", func() {
+		defer ticker.Stop()
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			case <-ticker.C:
+				r.evictStale(timeout)
+			}
+		}
+	})
+}
+
+func (r *Registry) evictStale(timeout time.Duration) {
+	r.mu.RLock()
+	var stale []string
+	now := time.Now()
+	for id, p := range r.providers {
+		if now.Sub(p.LastHeartbeat) > timeout {
+			stale = append(stale, id)
+		}
+	}
+	r.mu.RUnlock()
+
+	for _, id := range stale {
+		r.logger.Warn("evicting stale provider", "provider_id", id, "timeout", timeout)
+		r.Disconnect(id)
+	}
+}
