+++ coordinator/registry/persistence.go
@@ -1,0 +1,242 @@
+package registry
+
+import (
+	"context"
+	"encoding/json"
+	"time"
+
+	"github.com/eigeninference/d-inference/coordinator/protocol"
+	"github.com/eigeninference/d-inference/coordinator/saferun"
+	"github.com/eigeninference/d-inference/coordinator/store"
+)
+
+// LoadStoredProviders loads provider records and reputation from the store
+// on startup. This pre-populates a lookup table so that reconnecting providers
+// can have their trust level and reputation restored. Providers are NOT added
+// to the active registry (they need to reconnect via WebSocket first).
+func (r *Registry) LoadStoredProviders() map[string]*store.ProviderRecord {
+	if r.store == nil {
+		return nil
+	}
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	records, err := r.store.ListProviderRecords(ctx)
+	if err != nil {
+		r.logger.Warn("failed to load stored providers", "error", err)
+		return nil
+	}
+
+	lookup := make(map[string]*store.ProviderRecord, len(records))
+	for i := range records {
+		rec := records[i]
+		// Index by serial number for matching reconnecting providers
+		if rec.SerialNumber != "" {
+			lookup[rec.SerialNumber] = &rec
+		}
+		// Also index by SE public key
+		if rec.SEPublicKey != "" {
+			lookup["sekey:"+rec.SEPublicKey] = &rec
+		}
+	}
+
+	r.logger.Info("loaded stored provider records", "count", len(records))
+	return lookup
+}
+
+// RestoreProviderState restores trust level and reputation from a stored record
+// onto a live provider. Called after a provider reconnects and is matched to
+// its stored state by serial number or SE key.
+func (r *Registry) RestoreProviderState(p *Provider, rec *store.ProviderRecord) {
+	if rec == nil {
+		return
+	}
+
+	p.mu.Lock()
+	defer p.mu.Unlock()
+
+	// Restore trust level (will be re-verified via fresh attestation)
+	p.TrustLevel = TrustLevel(rec.TrustLevel)
+	p.Attested = rec.Attested
+	p.MDAVerified = rec.MDAVerified
+	p.ACMEVerified = rec.ACMEVerified
+
+	// Restore challenge state
+	if rec.LastChallengeVerified != nil {
+		p.LastChallengeVerified = *rec.LastChallengeVerified
+	}
+	p.FailedChallenges = rec.FailedChallenges
+
+	// Restore location only if the provider doesn't already have a fresh one
+	// (attachProviderLocation may have set it from the current request before
+	// RestoreProviderState runs).
+	if rec.Location != nil && p.Location == nil {
+		cp := *rec.Location
+		p.Location = &cp
+	}
+
+	// Restore account linkage
+	if rec.AccountID != "" && p.AccountID == "" {
+		p.AccountID = rec.AccountID
+	}
+
+	// Restore lifetime counters and the last raw session counters so future
+	// heartbeats can merge cleanly after coordinator or provider restarts.
+	p.Stats = protocol.HeartbeatStats{
+		RequestsServed:  rec.LifetimeRequestsServed,
+		TokensGenerated: rec.LifetimeTokensGenerated,
+	}
+	p.lastSessionStats = protocol.HeartbeatStats{
+		RequestsServed:  rec.LastSessionRequestsServed,
+		TokensGenerated: rec.LastSessionTokensGenerated,
+	}
+
+	// Restore reputation from store
+	if r.store != nil {
+		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		repRec, err := r.store.GetReputation(ctx, rec.ID)
+		if err == nil {
+			p.Reputation.TotalJobs = repRec.TotalJobs
+			p.Reputation.SuccessfulJobs = repRec.SuccessfulJobs
+			p.Reputation.FailedJobs = repRec.FailedJobs
+			p.Reputation.TotalUptime = time.Duration(repRec.TotalUptimeSeconds) * time.Second
+			p.Reputation.AvgResponseTime = time.Duration(repRec.AvgResponseTimeMs) * time.Millisecond
+			p.Reputation.ChallengesPassed = repRec.ChallengesPassed
+			p.Reputation.ChallengesFailed = repRec.ChallengesFailed
+		}
+	}
+
+	r.logger.Info("restored provider state from store",
+		"provider_id", p.ID,
+		"stored_id", rec.ID,
+		"trust_level", rec.TrustLevel,
+		"attested", rec.Attested,
+		"serial", rec.SerialNumber,
+	)
+}
+
+// PersistProvider unconditionally persists provider state to the store.
+// Use for critical state changes (attestation, trust level, disconnect).
+func (r *Registry) PersistProvider(p *Provider) {
+	r.persistProviderNow(p)
+}
+
+// PersistProviderThrottled persists provider state at most once per 30 seconds.
+// Use for high-frequency updates (heartbeats) that would otherwise saturate the
+// DB connection pool. Skipped writes are not lost — the next unthrottled persist
+// or the next throttle window will capture the current state.
+func (r *Registry) PersistProviderThrottled(p *Provider) {
+	const minInterval = 30 * time.Second
+	p.mu.Lock()
+	if time.Since(p.lastPersisted) < minInterval {
+		p.mu.Unlock()
+		return
+	}
+	p.lastPersisted = time.Now()
+	p.mu.Unlock()
+	r.persistProviderNow(p)
+}
+
+// persistProviderNow saves a provider's current state to the store.
+// Called asynchronously to avoid blocking the hot path.
+func (r *Registry) persistProviderNow(p *Provider) {
+	if r.store == nil {
+		return
+	}
+	saferun.Go(r.logger, "registry.persistProvider", func() {
+		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+
+		p.mu.Lock()
+		hardwareJSON, _ := json.Marshal(p.Hardware)
+		modelsJSON, _ := json.Marshal(p.Models)
+		var attestJSON json.RawMessage
+		if p.AttestationResult != nil {
+			attestJSON, _ = json.Marshal(p.AttestationResult)
+		}
+		seKey := ""
+		serial := ""
+		if p.AttestationResult != nil {
+			seKey = p.AttestationResult.PublicKey
+			serial = p.AttestationResult.SerialNumber
+		}
+		var mdaCertJSON json.RawMessage
+		if len(p.MDACertChain) > 0 {
+			mdaCertJSON, _ = json.Marshal(p.MDACertChain)
+		}
+		var lastChallenge *time.Time
+		if !p.LastChallengeVerified.IsZero() {
+			t := p.LastChallengeVerified
+			lastChallenge = &t
+		}
+
+		var locationCopy *store.ProviderLocation
+		if p.Location != nil {
+			lc := *p.Location
+			locationCopy = &lc
+		}
+
+		rec := store.ProviderRecord{
+			ID:                         p.ID,
+			Hardware:                   hardwareJSON,
+			Models:                     modelsJSON,
+			Backend:                    p.Backend,
+			Location:                   locationCopy,
+			TrustLevel:                 string(p.TrustLevel),
+			Attested:                   p.Attested,
+			AttestationResult:          attestJSON,
+			SEPublicKey:                seKey,
+			SerialNumber:               serial,
+			MDAVerified:                p.MDAVerified,
+			MDACertChain:               mdaCertJSON,
+			ACMEVerified:               p.ACMEVerified,
+			Version:                    p.Version,
+			RuntimeVerified:            p.RuntimeVerified,
+			PythonHash:                 p.PythonHash,
+			RuntimeHash:                p.RuntimeHash,
+			LastChallengeVerified:      lastChallenge,
+			FailedChallenges:           p.FailedChallenges,
+			AccountID:                  p.AccountID,
+			LifetimeRequestsServed:     p.Stats.RequestsServed,
+			LifetimeTokensGenerated:    p.Stats.TokensGenerated,
+			LastSessionRequestsServed:  p.lastSessionStats.RequestsServed,
+			LastSessionTokensGenerated: p.lastSessionStats.TokensGenerated,
+			RegisteredAt:               time.Now(),
+			LastSeen:                   time.Now(),
+		}
+		p.mu.Unlock()
+
+		if err := r.store.UpsertProvider(ctx, rec); err != nil {
+			r.logger.Warn("failed to persist provider", "provider_id", p.ID, "error", err)
+		}
+	})
+}
+
+// persistReputation saves a provider's current reputation to the store.
+// Called asynchronously to avoid blocking the hot path.
+func (r *Registry) persistReputation(p *Provider) {
+	if r.store == nil {
+		return
+	}
+	saferun.Go(r.logger, "registry.persistReputation", func() {
+		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+
+		p.mu.Lock()
+		rep := store.ReputationRecord{
+			TotalJobs:          p.Reputation.TotalJobs,
+			SuccessfulJobs:     p.Reputation.SuccessfulJobs,
+			FailedJobs:         p.Reputation.FailedJobs,
+			TotalUptimeSeconds: int64(p.Reputation.TotalUptime / time.Second),
+			AvgResponseTimeMs:  int64(p.Reputation.AvgResponseTime / time.Millisecond),
+			ChallengesPassed:   p.Reputation.ChallengesPassed,
+			ChallengesFailed:   p.Reputation.ChallengesFailed,
+		}
+		p.mu.Unlock()
+
+		if err := r.store.UpsertReputation(ctx, p.ID, rep); err != nil {
+			r.logger.Warn("failed to persist reputation", "provider_id", p.ID, "error", err)
+		}
+	})
+}
