+++ coordinator/registry/quick_capacity.go
@@ -1,0 +1,143 @@
+package registry
+
+import (
+	"time"
+)
+
+// QuickCapacityCheck performs a fast, read-only scan of the provider fleet to
+// determine whether any provider could serve a request for the given model
+// right now. It runs the same gates as the full routing path (status, trust,
+// runtime, privacy, challenge freshness, concurrency headroom, slot state,
+// free memory) but does NOT reserve capacity or create pending requests.
+//
+// Returns:
+//   - candidateCount: providers that passed ALL gates (could route right now)
+//   - capacityRejections: providers that serve the model and passed structural
+//     gates but were rejected for capacity reasons (full concurrency, no free
+//     memory, etc.)
+//
+// This is used for the pre-flight 429 check: if candidateCount == 0 &&
+// capacityRejections > 0, providers exist but are all at capacity (429).
+// If candidateCount == 0 && capacityRejections == 0, no provider serves
+// the model at all (404/503).
+func (r *Registry) QuickCapacityCheck(model string, estimatedPromptTokens, requestedMaxTokens int, allowedSerials ...string) (candidateCount, capacityRejections int) {
+	// Use a dummy PendingRequest with the caller's actual token estimates
+	// for the admission gate (freeMemoryAdmits).
+	if estimatedPromptTokens <= 0 {
+		estimatedPromptTokens = 500
+	}
+	if requestedMaxTokens <= 0 {
+		requestedMaxTokens = defaultRequestedMaxTokens
+	}
+	dummyPR := &PendingRequest{
+		RequestID:             "capacity-check",
+		Model:                 model,
+		EstimatedPromptTokens: estimatedPromptTokens,
+		RequestedMaxTokens:    requestedMaxTokens,
+	}
+
+	// Build allowed serial set for optional provider filtering.
+	allowedSet := make(map[string]struct{}, len(allowedSerials))
+	for _, s := range allowedSerials {
+		allowedSet[s] = struct{}{}
+	}
+
+	r.mu.RLock()
+	defer r.mu.RUnlock()
+
+	now := time.Now()
+	for _, p := range r.providers {
+		// Filter by allowed serials before acquiring the provider lock
+		// (providerMatchesAllowedSerial takes p.mu internally).
+		if len(allowedSet) > 0 && !providerMatchesAllowedSerial(p, allowedSet) {
+			continue
+		}
+
+		p.mu.Lock()
+
+		// Structural gates (same as snapshotProviderLocked).
+		if !r.providerServesCatalogModelLocked(p, model) {
+			p.mu.Unlock()
+			continue
+		}
+		if p.Status == StatusOffline || p.Status == StatusUntrusted {
+			p.mu.Unlock()
+			continue
+		}
+		if trustRank(p.TrustLevel) < trustRank(r.MinTrustLevel) {
+			p.mu.Unlock()
+			continue
+		}
+		if !p.RuntimeVerified {
+			p.mu.Unlock()
+			continue
+		}
+		if !providerSupportsPrivateTextLocked(p) {
+			p.mu.Unlock()
+			continue
+		}
+		if p.LastChallengeVerified.IsZero() || now.Sub(p.LastChallengeVerified) > challengeFreshnessMaxAge {
+			p.mu.Unlock()
+			continue
+		}
+
+		// Concurrency gate.
+		if !p.hasConcurrencyHeadroomForModelLocked(model) {
+			p.mu.Unlock()
+			capacityRejections++
+			continue
+		}
+
+		// Build a snapshot for the admission gate (slot state + free memory).
+		snap := routingSnapshot{
+			provider:      p,
+			model:         model,
+			slotState:     "unknown",
+			totalPending:  p.pendingCount(),
+			totalMemoryGB: float64(p.Hardware.MemoryGB),
+			modelSizeGB:   r.catalogSizeGBLocked(model),
+		}
+		for _, pending := range p.pendingReqs {
+			if pending.Model != model {
+				continue
+			}
+			snap.pendingForModel++
+			snap.pendingMaxTokens += pendingTokenBudget(pending)
+		}
+		if p.BackendCapacity != nil {
+			snap.gpuMemoryActiveGB = p.BackendCapacity.GPUMemoryActiveGB
+			if p.BackendCapacity.TotalMemoryGB > 0 {
+				snap.totalMemoryGB = p.BackendCapacity.TotalMemoryGB
+			}
+			for _, slot := range p.BackendCapacity.Slots {
+				if slot.Model != model {
+					continue
+				}
+				snap.slotState = slot.State
+				snap.activeTokenBudgetUsed = slot.ActiveTokenBudgetUsed
+				snap.activeTokenBudgetMax = slot.ActiveTokenBudgetMax
+				snap.queuedTokenBudget = slot.QueuedTokenBudget
+				snap.maxTokensPotential = slot.MaxTokensPotential
+				break
+			}
+		}
+		snap.modelLoaded = snap.slotState == "running"
+		snap.availableOnDisk = !snap.modelLoaded
+
+		p.mu.Unlock()
+
+		// Slot state gate (crashed/reloading are ineligible).
+		if _, eligible := slotStatePenalty(snap.slotState); !eligible {
+			continue
+		}
+
+		// Free memory / token budget admission gate.
+		if !freeMemoryAdmits(snap, dummyPR.EstimatedPromptTokens, dummyPR.RequestedMaxTokens) {
+			capacityRejections++
+			continue
+		}
+
+		candidateCount++
+	}
+	return candidateCount, capacityRejections
+}
