+++ coordinator/registry/registry.go
@@ -18,496 +18,13 @@
 package registry
 
 import (
-	"context"
-	"encoding/base64"
-	"encoding/json"
-	"fmt"
 	"log/slog"
-	"math"
-	"math/rand"
-	"sort"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"github.com/eigeninference/d-inference/coordinator/attestation"
-	"github.com/eigeninference/d-inference/coordinator/protocol"
-	"github.com/eigeninference/d-inference/coordinator/saferun"
 	"github.com/eigeninference/d-inference/coordinator/store"
-	"nhooyr.io/websocket"
 )
-
-// ProviderStatus represents the operational state of a provider.
-type ProviderStatus string
-
-const (
-	StatusOnline    ProviderStatus = "online"
-	StatusServing   ProviderStatus = "serving"
-	StatusOffline   ProviderStatus = "offline"
-	StatusUntrusted ProviderStatus = "untrusted"
-)
-
-// TrustLevel represents the attestation trust level of a provider.
-type TrustLevel string
-
-const (
-	TrustNone       TrustLevel = "none"        // No attestation provided
-	TrustSelfSigned TrustLevel = "self_signed" // Attestation signed by provider's own key
-	TrustHardware   TrustLevel = "hardware"    // MDM + MDA + SE key bound to Apple-verified hardware
-)
-
-const BackendMLXSwift = "mlx-swift"
-
-// MaxFailedChallenges is the number of consecutive challenge failures before
-// a provider is marked untrusted and fully derouted.
-const MaxFailedChallenges = 3
-
-func BackendUsesSwiftRuntime(backend string) bool {
-	return backend == BackendMLXSwift
-}
-
-// PendingRequest is a channel-based handle for an in-flight inference request.
-type PendingRequest struct {
-	RequestID   string
-	ProviderID  string
-	Model       string
-	ConsumerKey string
-	// KeyID is the public ID of the API key that originated the request, used
-	// for per-key usage and spend attribution. Empty for account-scoped/legacy
-	// callers (Privy JWT, admin, provider tokens, unlinked keys without an ID).
-	KeyID string
-	// KeyLimitMicroUSD / KeyLimitReset carry the originating key's spend cap so
-	// the per-key cap can be re-enforced when a provider's custom price tops up
-	// the reservation above the platform rate. Nil limit = no per-key cap.
-	KeyLimitMicroUSD *int64
-	KeyLimitReset    string
-	ConsumerLocation *store.ProviderLocation
-	// IsResponsesAPI tracks requests received through /v1/responses so the
-	// coordinator can translate provider chat-completions output back into
-	// Responses API objects for SDK clients.
-	IsResponsesAPI bool
-	// AllowedProviderSerials optionally restricts routing to providers with
-	// one of these attested hardware serials. Empty means the request may
-	// route to any eligible provider.
-	AllowedProviderSerials []string
-	// EstimatedPromptTokens is a coordinator-side heuristic used only for
-	// routing and queue admission. It does not need tokenizer-perfect accuracy.
-	EstimatedPromptTokens int
-	// RequestedMaxTokens is the consumer's requested output budget (or a
-	// sensible default when omitted). It is used for backlog estimation.
-	RequestedMaxTokens int
-	AcceptedCh         chan struct{}           // signalled when provider accepts request
-	ChunkCh            chan string             // SSE data chunks
-	CompleteCh         chan protocol.UsageInfo // closed after usage sent
-	ErrorCh            chan protocol.InferenceErrorMessage
-	SessionPrivKey     *[32]byte // E2E session private key for decrypting responses
-	SESignature        string    // SE signature over response hash
-	ResponseHash       string    // SHA-256 of response data
-
-	// ReservedMicroUSD is the balance atomically debited at pre-flight.
-	// The post-inference charge adjusts for the difference between the
-	// actual cost and this reservation, preventing billing race conditions.
-	ReservedMicroUSD int64
-	// BaseReservedMicroUSD is the shared base reservation (platform price)
-	// charged once per request. ReservedMicroUSD may exceed it after a
-	// provider-specific top-up; the difference (the per-attempt "extra") must
-	// be refunded if this attempt is abandoned (speculative loser, retry,
-	// timeout). The base itself is refunded once globally or settled by the
-	// winning attempt.
-	BaseReservedMicroUSD int64
-	reservationMu        sync.Mutex
-	reservationFinalized bool
-
-	// Timing fields for latency decomposition.
-	Timing *RequestTiming
-}
-
-// MarkReservationFinalized returns true only for the first settlement or refund
-// of a pre-flight balance reservation. It prevents a terminal provider error
-// racing with a late completion from crediting or refunding the same reservation
-// twice.
-func (pr *PendingRequest) MarkReservationFinalized() bool {
-	ok, _ := pr.FinalizeReservation(nil)
-	return ok
-}
-
-// FinalizeReservation runs settle while holding the reservation finalization
-// lock and marks the reservation finalized only if settle succeeds. It returns
-// false when another terminal path already finalized the reservation.
-func (pr *PendingRequest) FinalizeReservation(settle func() error) (bool, error) {
-	pr.reservationMu.Lock()
-	defer pr.reservationMu.Unlock()
-	if pr.reservationFinalized {
-		return false, nil
-	}
-	if settle != nil {
-		if err := settle(); err != nil {
-			return false, err
-		}
-	}
-	pr.reservationFinalized = true
-	return true, nil
-}
-
-type RequestTiming struct {
-	ReceivedAt   time.Time // handler entry
-	ParsedAt     time.Time // after parse + validate
-	ReservedAt   time.Time // after balance reservation
-	RoutedAt     time.Time // after provider selection (including queue wait)
-	EncryptedAt  time.Time // after E2E encryption
-	QueuedAt     time.Time // set when request enters the queue
-	DispatchedAt time.Time // set when request is sent to provider via WebSocket
-	FirstChunkAt time.Time // set when first inference chunk arrives from provider
-}
-
-// Provider represents a connected provider agent.
-type Provider struct {
-	ID                string
-	Hardware          protocol.Hardware
-	Models            []protocol.ModelInfo
-	Backend           string
-	Location          *store.ProviderLocation
-	PublicKey         string // base64-encoded X25519 public key for E2E encryption
-	Attested          bool   // true if attestation was verified successfully
-	AttestationResult *attestation.VerificationResult
-	TrustLevel        TrustLevel             // attestation trust level
-	MDAVerified       bool                   // true if Apple Device Attestation cert chain verified
-	MDACertChain      [][]byte               // DER-encoded Apple MDA certificate chain (leaf first)
-	MDAResult         *attestation.MDAResult // parsed OIDs from Apple cert
-	ACMEVerified      bool                   // true if ACME device-attest-01 client cert verified (SE key proven)
-	SEKeyBound        bool                   // true if SE key was bound to device via MDA nonce
-	Status            ProviderStatus
-	Conn              *websocket.Conn
-	LastHeartbeat     time.Time
-	Stats             protocol.HeartbeatStats // lifetime counters shown to users
-	lastSessionStats  protocol.HeartbeatStats // raw counters from the current provider process
-
-	// Account linkage (set when provider authenticates via device auth token)
-	AccountID string // internal account ID (from device auth flow)
-
-	// Benchmark data reported at registration
-	PrefillTPS float64 // prefill tokens per second
-	DecodeTPS  float64 // decode tokens per second
-
-	// Warm model cache tracking
-	WarmModels   []string // models currently loaded in provider's memory
-	CurrentModel string   // model currently being served
-
-	// Live system metrics from heartbeats
-	SystemMetrics protocol.SystemMetrics
-
-	// Live backend capacity from heartbeats (nil for providers without capacity reporting)
-	BackendCapacity *protocol.BackendCapacity
-
-	// Reputation tracking
-	Reputation Reputation
-
-	// Version and runtime integrity verification
-	Version                 string `json:"version,omitempty"`                   // provider binary version (e.g. "0.2.31")
-	RuntimeVerified         bool   `json:"runtime_verified"`                    // true if runtime hashes match the known-good manifest
-	RuntimeManifestChecked  bool   `json:"runtime_manifest_checked"`            // true only when a manifest was present and hashes were verified (fail-closed for text)
-	EncryptedResponseChunks bool   `json:"encrypted_response_chunks,omitempty"` // true when text response chunks are encrypted to the coordinator
-	PythonHash              string `json:"python_hash,omitempty"`
-	RuntimeHash             string `json:"runtime_hash,omitempty"`
-	TemplateHashes          map[string]string
-
-	// Phase 7: Privacy invariant attestation.
-	// Self-reported by the provider at registration. SIPEnabled is overridden
-	// by the coordinator after each attestation challenge response with a
-	// coordinator-verified value. HypervisorActive is informational.
-	PrivacyCapabilities *protocol.PrivacyCapabilities `json:"privacy_capabilities,omitempty"`
-
-	// Coordinator-verified SIP status from the most recent attestation challenge.
-	// Unlike PrivacyCapabilities.SIPEnabled (provider self-report at registration),
-	// this is set by the coordinator after independently checking the challenge response.
-	ChallengeVerifiedSIP bool `json:"challenge_verified_sip"`
-
-	// lastPersisted tracks when this provider was last written to the store.
-	// Used by PersistProviderThrottled to avoid hammering Postgres on every heartbeat.
-	lastPersisted time.Time
-
-	// Challenge-response verification state
-	LastChallengeVerified time.Time // last successful challenge verification
-	FailedChallenges      int       // consecutive failed challenges
-
-	// untrustedRecoverable marks an untrust as a *transient* missed-challenge
-	// deroute (timeout / no-response) that may self-recover on the next passing
-	// challenge. It is false for every hard/security deroute. In-memory only —
-	// never persisted, because recoverability is meaningless without a live
-	// WebSocket and a running challenge loop.
-	untrustedRecoverable bool
-
-	mu          sync.Mutex
-	pendingReqs map[string]*PendingRequest
-}
-
-func providerSupportsPrivateTextLocked(p *Provider) bool {
-	if p.PublicKey == "" || !privateTextBackendSupported(p.Backend) || !p.EncryptedResponseChunks {
-		return false
-	}
-	if !p.RuntimeManifestChecked {
-		return false
-	}
-	// Require coordinator-verified SIP (from attestation challenge) rather
-	// than trusting the provider's self-reported SIPEnabled field.
-	if !p.ChallengeVerifiedSIP {
-		return false
-	}
-	caps := p.PrivacyCapabilities
-	if caps == nil {
-		return false
-	}
-	// Only mlx-swift is routable (enforced by privateTextBackendSupported above).
-	// Python-specific caps (PythonRuntimeLocked, DangerousModulesBlocked) are
-	// retained in the protocol struct for wire backward compat but are no longer
-	// required for routing.
-	return caps.TextBackendInprocess &&
-		caps.TextProxyDisabled &&
-		caps.AntiDebugEnabled &&
-		caps.CoreDumpsDisabled &&
-		caps.EnvScrubbed
-}
-
-func privateTextBackendSupported(backend string) bool {
-	// Python/legacy inprocess-mlx backend is deprecated and no longer
-	// routable. Only Swift (mlx-swift) providers are admitted.
-	return backend == BackendMLXSwift
-}
-
-// AddPending registers a pending request on this provider.
-func (p *Provider) AddPending(pr *PendingRequest) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.addPendingLocked(pr)
-}
-
-// addPendingLocked registers a pending request. Caller must hold p.mu.
-func (p *Provider) addPendingLocked(pr *PendingRequest) {
-	p.pendingReqs[pr.RequestID] = pr
-}
-
-// RemovePending removes and returns a pending request.
-func (p *Provider) RemovePending(requestID string) *PendingRequest {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.removePendingLocked(requestID)
-}
-
-// removePendingLocked removes and returns a pending request. Caller must hold p.mu.
-func (p *Provider) removePendingLocked(requestID string) *PendingRequest {
-	pr := p.pendingReqs[requestID]
-	delete(p.pendingReqs, requestID)
-	return pr
-}
-
-// GetPending retrieves a pending request without removing it.
-func (p *Provider) GetPending(requestID string) *PendingRequest {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.pendingReqs[requestID]
-}
-
-// SetAttested updates attestation state (thread-safe).
-// Note: persistence is handled by the Registry methods that call this,
-// via persistProvider() after attestation verification completes.
-func (p *Provider) SetAttested(attested bool, trust TrustLevel) {
-	p.mu.Lock()
-	p.Attested = attested
-	p.TrustLevel = trust
-	p.mu.Unlock()
-}
-
-// SetLastChallengeVerified updates the challenge timestamp (thread-safe).
-func (p *Provider) SetLastChallengeVerified(t time.Time) {
-	p.mu.Lock()
-	p.LastChallengeVerified = t
-	p.mu.Unlock()
-}
-
-// GetLastChallengeVerified returns the last challenge verification time (thread-safe).
-func (p *Provider) GetLastChallengeVerified() time.Time {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.LastChallengeVerified
-}
-
-// GetChallengeVerifiedSIP returns whether SIP was verified in the last challenge (thread-safe).
-func (p *Provider) GetChallengeVerifiedSIP() bool {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.ChallengeVerifiedSIP
-}
-
-func (p *Provider) SetChallengeVerifiedSIP(v bool) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	p.ChallengeVerifiedSIP = v
-}
-
-// Mu returns the provider's mutex for external callers that need to read
-// fields like Status atomically. Prefer dedicated getters where available.
-func (p *Provider) Mu() *sync.Mutex {
-	return &p.mu
-}
-
-// ChallengeShouldStop reports whether the attestation challenge loop should
-// stop for this provider. It stops only for a *hard* (non-recoverable) untrust;
-// a transiently-untrusted provider keeps being challenged so a later passing
-// challenge can restore it via RecordChallengeSuccess. Thread-safe.
-func (p *Provider) ChallengeShouldStop() bool {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.Status == StatusUntrusted && !p.untrustedRecoverable
-}
-
-// SetAttestationResult stores the parsed attestation result (thread-safe).
-func (p *Provider) SetAttestationResult(result *attestation.VerificationResult) {
-	p.mu.Lock()
-	p.AttestationResult = result
-	p.mu.Unlock()
-}
-
-// GetAttestationResult returns the current attestation result (thread-safe).
-func (p *Provider) GetAttestationResult() *attestation.VerificationResult {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.AttestationResult
-}
-
-// pendingCount returns the number of in-flight requests.
-// Caller must hold p.mu.
-func (p *Provider) pendingCount() int {
-	return len(p.pendingReqs)
-}
-
-// PendingCount returns the number of in-flight requests (thread-safe).
-func (p *Provider) PendingCount() int {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.pendingCount()
-}
-
-// MaxConcurrency returns the dynamic max concurrent request limit.
-// Uses hardware-based estimation when backend capacity is reported.
-// Falls back to DefaultMaxConcurrent for providers without capacity reporting.
-func (p *Provider) MaxConcurrency() int {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.maxConcurrency()
-}
-
-// MaxConcurrencyForModel returns the concurrency limit for a specific model.
-// A positive provider-reported slot cap wins; zero/missing preserves the
-// legacy provider-level fallback.
-func (p *Provider) MaxConcurrencyForModel(model string) int {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	return p.maxConcurrencyForModelLocked(model)
-}
-
-// maxConcurrency is the lock-free version (caller must hold p.mu).
-//
-// Tier values were lowered in Phase 2 of the routing-algorithm rework
-// (was 4/8/16/24/32). The old caps were derived from "how many
-// requests can theoretically fit in GPU memory"; the new caps reflect
-// "how many concurrent decodes a single MLX backend can run before
-// per-request TPS collapses". Empirically this is much smaller than
-// the memory-derived ceiling. Pushing past it makes each request slow
-// without increasing fleet throughput.
-func (p *Provider) maxConcurrency() int {
-	if p.BackendCapacity == nil {
-		return DefaultMaxConcurrent
-	}
-
-	// Token-budget providers use budget-based admission; the concurrency
-	// cap is just a safety valve.
-	for _, slot := range p.BackendCapacity.Slots {
-		if slot.ActiveTokenBudgetMax > 0 {
-			return 24
-		}
-	}
-
-	// Hardware-based cap using total memory reported by the provider.
-	memGB := p.BackendCapacity.TotalMemoryGB
-	if memGB <= 0 {
-		memGB = float64(p.Hardware.MemoryGB)
-	}
-	var cap int
-	switch {
-	case memGB <= 24:
-		cap = 2
-	case memGB <= 48:
-		cap = 4
-	case memGB <= 96:
-		cap = 6
-	case memGB <= 128:
-		cap = 8
-	default:
-		cap = 12
-	}
-	return cap
-}
-
-// maxConcurrencyForModelLocked is the lock-free model-aware concurrency cap.
-// Caller must hold p.mu.
-func (p *Provider) maxConcurrencyForModelLocked(model string) int {
-	if p.BackendCapacity != nil {
-		for _, slot := range p.BackendCapacity.Slots {
-			if slot.Model == model && slot.MaxConcurrency > 0 {
-				return slot.MaxConcurrency
-			}
-		}
-	}
-	return p.maxConcurrency()
-}
-
-func (p *Provider) pendingCountForModelLocked(model string) int {
-	count := 0
-	for _, pr := range p.pendingReqs {
-		if pr.Model == model {
-			count++
-		}
-	}
-	return count
-}
-
-func (p *Provider) hasReportedMaxConcurrencyForModelLocked(model string) bool {
-	if p.BackendCapacity == nil {
-		return false
-	}
-	for _, slot := range p.BackendCapacity.Slots {
-		if slot.Model == model && slot.MaxConcurrency > 0 {
-			return true
-		}
-	}
-	return false
-}
-
-func (p *Provider) pendingLoadForModelLocked(model string) int {
-	if !p.hasReportedMaxConcurrencyForModelLocked(model) {
-		return p.pendingCount()
-	}
-	load := p.pendingCountForModelLocked(model)
-	if p.BackendCapacity != nil {
-		for _, slot := range p.BackendCapacity.Slots {
-			if slot.Model != model {
-				continue
-			}
-			backendLoad := slot.NumRunning + slot.NumWaiting
-			if backendLoad > load {
-				load = backendLoad
-			}
-			break
-		}
-	}
-	return load
-}
-
-func (p *Provider) hasConcurrencyHeadroomForModelLocked(model string) bool {
-	return p.pendingLoadForModelLocked(model) < p.maxConcurrencyForModelLocked(model) &&
-		p.pendingCount() < p.maxConcurrency()
-}
 
 // Registry holds all connected providers and provides routing.
 type Registry struct {
@@ -562,355 +79,6 @@ func (r *Registry) SetStore(st store.Store) {
 	r.store = st
 }
 
-// LoadStoredProviders loads provider records and reputation from the store
-// on startup. This pre-populates a lookup table so that reconnecting providers
-// can have their trust level and reputation restored. Providers are NOT added
-// to the active registry (they need to reconnect via WebSocket first).
-func (r *Registry) LoadStoredProviders() map[string]*store.ProviderRecord {
-	if r.store == nil {
-		return nil
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-
-	records, err := r.store.ListProviderRecords(ctx)
-	if err != nil {
-		r.logger.Warn("failed to load stored providers", "error", err)
-		return nil
-	}
-
-	lookup := make(map[string]*store.ProviderRecord, len(records))
-	for i := range records {
-		rec := records[i]
-		// Index by serial number for matching reconnecting providers
-		if rec.SerialNumber != "" {
-			lookup[rec.SerialNumber] = &rec
-		}
-		// Also index by SE public key
-		if rec.SEPublicKey != "" {
-			lookup["sekey:"+rec.SEPublicKey] = &rec
-		}
-	}
-
-	r.logger.Info("loaded stored provider records", "count", len(records))
-	return lookup
-}
-
-// RestoreProviderState restores trust level and reputation from a stored record
-// onto a live provider. Called after a provider reconnects and is matched to
-// its stored state by serial number or SE key.
-func (r *Registry) RestoreProviderState(p *Provider, rec *store.ProviderRecord) {
-	if rec == nil {
-		return
-	}
-
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	// Restore trust level (will be re-verified via fresh attestation)
-	p.TrustLevel = TrustLevel(rec.TrustLevel)
-	p.Attested = rec.Attested
-	p.MDAVerified = rec.MDAVerified
-	p.ACMEVerified = rec.ACMEVerified
-
-	// Restore challenge state
-	if rec.LastChallengeVerified != nil {
-		p.LastChallengeVerified = *rec.LastChallengeVerified
-	}
-	p.FailedChallenges = rec.FailedChallenges
-
-	// Restore location only if the provider doesn't already have a fresh one
-	// (attachProviderLocation may have set it from the current request before
-	// RestoreProviderState runs).
-	if rec.Location != nil && p.Location == nil {
-		cp := *rec.Location
-		p.Location = &cp
-	}
-
-	// Restore account linkage
-	if rec.AccountID != "" && p.AccountID == "" {
-		p.AccountID = rec.AccountID
-	}
-
-	// Restore lifetime counters and the last raw session counters so future
-	// heartbeats can merge cleanly after coordinator or provider restarts.
-	p.Stats = protocol.HeartbeatStats{
-		RequestsServed:  rec.LifetimeRequestsServed,
-		TokensGenerated: rec.LifetimeTokensGenerated,
-	}
-	p.lastSessionStats = protocol.HeartbeatStats{
-		RequestsServed:  rec.LastSessionRequestsServed,
-		TokensGenerated: rec.LastSessionTokensGenerated,
-	}
-
-	// Restore reputation from store
-	if r.store != nil {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		repRec, err := r.store.GetReputation(ctx, rec.ID)
-		if err == nil {
-			p.Reputation.TotalJobs = repRec.TotalJobs
-			p.Reputation.SuccessfulJobs = repRec.SuccessfulJobs
-			p.Reputation.FailedJobs = repRec.FailedJobs
-			p.Reputation.TotalUptime = time.Duration(repRec.TotalUptimeSeconds) * time.Second
-			p.Reputation.AvgResponseTime = time.Duration(repRec.AvgResponseTimeMs) * time.Millisecond
-			p.Reputation.ChallengesPassed = repRec.ChallengesPassed
-			p.Reputation.ChallengesFailed = repRec.ChallengesFailed
-		}
-	}
-
-	r.logger.Info("restored provider state from store",
-		"provider_id", p.ID,
-		"stored_id", rec.ID,
-		"trust_level", rec.TrustLevel,
-		"attested", rec.Attested,
-		"serial", rec.SerialNumber,
-	)
-}
-
-// PersistProvider unconditionally persists provider state to the store.
-// Use for critical state changes (attestation, trust level, disconnect).
-func (r *Registry) PersistProvider(p *Provider) {
-	r.persistProviderNow(p)
-}
-
-// PersistProviderThrottled persists provider state at most once per 30 seconds.
-// Use for high-frequency updates (heartbeats) that would otherwise saturate the
-// DB connection pool. Skipped writes are not lost — the next unthrottled persist
-// or the next throttle window will capture the current state.
-func (r *Registry) PersistProviderThrottled(p *Provider) {
-	const minInterval = 30 * time.Second
-	p.mu.Lock()
-	if time.Since(p.lastPersisted) < minInterval {
-		p.mu.Unlock()
-		return
-	}
-	p.lastPersisted = time.Now()
-	p.mu.Unlock()
-	r.persistProviderNow(p)
-}
-
-// persistProviderNow saves a provider's current state to the store.
-// Called asynchronously to avoid blocking the hot path.
-func (r *Registry) persistProviderNow(p *Provider) {
-	if r.store == nil {
-		return
-	}
-	saferun.Go(r.logger, "registry.persistProvider", func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		p.mu.Lock()
-		hardwareJSON, _ := json.Marshal(p.Hardware)
-		modelsJSON, _ := json.Marshal(p.Models)
-		var attestJSON json.RawMessage
-		if p.AttestationResult != nil {
-			attestJSON, _ = json.Marshal(p.AttestationResult)
-		}
-		seKey := ""
-		serial := ""
-		if p.AttestationResult != nil {
-			seKey = p.AttestationResult.PublicKey
-			serial = p.AttestationResult.SerialNumber
-		}
-		var mdaCertJSON json.RawMessage
-		if len(p.MDACertChain) > 0 {
-			mdaCertJSON, _ = json.Marshal(p.MDACertChain)
-		}
-		var lastChallenge *time.Time
-		if !p.LastChallengeVerified.IsZero() {
-			t := p.LastChallengeVerified
-			lastChallenge = &t
-		}
-
-		var locationCopy *store.ProviderLocation
-		if p.Location != nil {
-			lc := *p.Location
-			locationCopy = &lc
-		}
-
-		rec := store.ProviderRecord{
-			ID:                         p.ID,
-			Hardware:                   hardwareJSON,
-			Models:                     modelsJSON,
-			Backend:                    p.Backend,
-			Location:                   locationCopy,
-			TrustLevel:                 string(p.TrustLevel),
-			Attested:                   p.Attested,
-			AttestationResult:          attestJSON,
-			SEPublicKey:                seKey,
-			SerialNumber:               serial,
-			MDAVerified:                p.MDAVerified,
-			MDACertChain:               mdaCertJSON,
-			ACMEVerified:               p.ACMEVerified,
-			Version:                    p.Version,
-			RuntimeVerified:            p.RuntimeVerified,
-			PythonHash:                 p.PythonHash,
-			RuntimeHash:                p.RuntimeHash,
-			LastChallengeVerified:      lastChallenge,
-			FailedChallenges:           p.FailedChallenges,
-			AccountID:                  p.AccountID,
-			LifetimeRequestsServed:     p.Stats.RequestsServed,
-			LifetimeTokensGenerated:    p.Stats.TokensGenerated,
-			LastSessionRequestsServed:  p.lastSessionStats.RequestsServed,
-			LastSessionTokensGenerated: p.lastSessionStats.TokensGenerated,
-			RegisteredAt:               time.Now(),
-			LastSeen:                   time.Now(),
-		}
-		p.mu.Unlock()
-
-		if err := r.store.UpsertProvider(ctx, rec); err != nil {
-			r.logger.Warn("failed to persist provider", "provider_id", p.ID, "error", err)
-		}
-	})
-}
-
-// persistReputation saves a provider's current reputation to the store.
-// Called asynchronously to avoid blocking the hot path.
-func (r *Registry) persistReputation(p *Provider) {
-	if r.store == nil {
-		return
-	}
-	saferun.Go(r.logger, "registry.persistReputation", func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		p.mu.Lock()
-		rep := store.ReputationRecord{
-			TotalJobs:          p.Reputation.TotalJobs,
-			SuccessfulJobs:     p.Reputation.SuccessfulJobs,
-			FailedJobs:         p.Reputation.FailedJobs,
-			TotalUptimeSeconds: int64(p.Reputation.TotalUptime / time.Second),
-			AvgResponseTimeMs:  int64(p.Reputation.AvgResponseTime / time.Millisecond),
-			ChallengesPassed:   p.Reputation.ChallengesPassed,
-			ChallengesFailed:   p.Reputation.ChallengesFailed,
-		}
-		p.mu.Unlock()
-
-		if err := r.store.UpsertReputation(ctx, p.ID, rep); err != nil {
-			r.logger.Warn("failed to persist reputation", "provider_id", p.ID, "error", err)
-		}
-	})
-}
-
-// TruncHash returns the first 16 chars of a hash string for logging.
-func TruncHash(h string) string {
-	if len(h) > 16 {
-		return h[:16] + "..."
-	}
-	return h
-}
-
-// CatalogEntry holds metadata about an active model in the catalog.
-type CatalogEntry struct {
-	ID         string
-	WeightHash string  // expected SHA-256 weight fingerprint (empty = not enforced)
-	SizeGB     float64 // disk/GPU footprint of the model weights (zero = unknown, gate disabled)
-}
-
-// SetModelCatalog updates the set of active models. Only models in this
-// set will be accepted from providers during registration and routable to
-// consumers. Pass nil to disable catalog filtering for tests/dev flows. Passing
-// an empty non-nil slice configures a deny-all catalog, which is what a fresh
-// DB-backed registry should do until an operator registers and promotes models.
-func (r *Registry) SetModelCatalog(entries []CatalogEntry) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if entries == nil {
-		r.modelCatalog = nil
-		return
-	}
-	catalog := make(map[string]CatalogEntry, len(entries))
-	for _, e := range entries {
-		catalog[e.ID] = e
-	}
-	r.modelCatalog = catalog
-}
-
-// ModelType returns the model type string for the given model ID, or
-// "unknown" if no provider is currently serving it.
-func (r *Registry) ModelType(model string) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	for _, p := range r.providers {
-		p.mu.Lock()
-		for _, m := range p.Models {
-			if m.ID == model && m.ModelType != "" {
-				p.mu.Unlock()
-				return m.ModelType
-			}
-		}
-		p.mu.Unlock()
-	}
-	return "unknown"
-}
-
-// IsModelInCatalog returns true if the model is in the active catalog, or if
-// catalog filtering has been explicitly disabled by setting a nil catalog.
-func (r *Registry) IsModelInCatalog(model string) bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	if r.modelCatalog == nil {
-		return true
-	}
-	_, ok := r.modelCatalog[model]
-	return ok
-}
-
-// CatalogWeightHash returns the expected weight hash for a model, or empty
-// string if not set or not in catalog.
-func (r *Registry) CatalogWeightHash(model string) string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	if e, ok := r.modelCatalog[model]; ok {
-		return e.WeightHash
-	}
-	return ""
-}
-
-// modelAllowedByCatalogLocked returns whether a provider-reported model is
-// allowed by the current catalog. Caller must hold r.mu (read or write). A nil
-// catalog disables filtering; an empty non-nil catalog denies all models.
-func (r *Registry) modelAllowedByCatalogLocked(model protocol.ModelInfo) bool {
-	if r.modelCatalog == nil {
-		return true
-	}
-	entry, ok := r.modelCatalog[model.ID]
-	if !ok {
-		return false
-	}
-	return entry.WeightHash == "" || model.WeightHash == "" || model.WeightHash == entry.WeightHash
-}
-
-// providerServesCatalogModelLocked returns true if the provider advertises the
-// model and that model is currently allowed by the catalog. Caller must hold
-// r.mu and p.mu.
-func (r *Registry) providerServesCatalogModelLocked(p *Provider, model string) bool {
-	for _, m := range p.Models {
-		if m.ID == model && r.modelAllowedByCatalogLocked(m) {
-			return true
-		}
-	}
-	return false
-}
-
-// catalogSizeGBLocked returns the model's reported weight footprint in GB,
-// or 0 when unknown. Caller must hold r.mu (read or write). Zero means the
-// memory-admission gate should not enforce for this model — typically a
-// catalog entry that pre-dates the SizeGB field, or a model the operator
-// hasn't sized yet.
-func (r *Registry) catalogSizeGBLocked(model string) float64 {
-	if e, ok := r.modelCatalog[model]; ok {
-		return e.SizeGB
-	}
-	return 0
-}
-
-// trustMeetsMinimum returns true if the given trust level meets the minimum.
-func (r *Registry) trustMeetsMinimum(level TrustLevel) bool {
-	return trustRank(level) >= trustRank(r.MinTrustLevel)
-}
-
 // Queue returns the registry's request queue.
 func (r *Registry) Queue() *RequestQueue {
 	return r.queue
@@ -922,1888 +90,4 @@ func (r *Registry) SetQueue(q *RequestQueue) {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	r.queue = q
-}
-
-// Sanity caps on provider-reported stats. A malicious (or broken) provider
-// could otherwise report absurd values to monopolize routing. These caps are
-// ~3-4x current hardware ceilings (M2 Ultra is ~800 GB/s, MLX decode is ~120
-// tok/s, max Mac Studio RAM is 512 GB) so legitimate future hardware isn't
-// clamped unnecessarily.
-const (
-	maxDecodeTPS                    = 500.0
-	maxPrefillTPS                   = 5000.0
-	maxMemoryBandwidthGBs           = 2000.0
-	maxMemoryGB                     = 1024
-	maxMemoryGBFloat                = 1024.0
-	maxReportedMaxConcurrency       = 24
-	maxTokensPotential              = 1_000_000
-	maxTokenBudgetCap         int64 = 10_000_000_000 // 10 billion — generous safety valve for total token budget capacity
-)
-
-// clampNonNeg returns v clamped into [0, max]; NaN/negative become 0.
-// The bool is true if the value was out of range.
-func clampNonNeg(v, max float64) (float64, bool) {
-	if math.IsNaN(v) || v < 0 {
-		return 0, true
-	}
-	if v > max {
-		return max, true
-	}
-	return v, false
-}
-
-// clampBackendCapacity applies sanity caps to provider-reported backend
-// capacity fields that feed the routing scorer. A provider reporting
-// TotalMemoryGB=1e9 would make gpuUtil ~= 0 and dodge health penalties, so
-// we cap it at maxMemoryGBFloat. Same for MaxTokensPotential which directly
-// controls backlog cost. NaN/negative become 0.
-func clampBackendCapacity(logger *slog.Logger, providerID string, bc *protocol.BackendCapacity) {
-	if bc == nil {
-		return
-	}
-	if v, changed := clampNonNeg(bc.TotalMemoryGB, maxMemoryGBFloat); changed {
-		logger.Warn("provider total_memory_gb out of range, clamping",
-			"provider_id", providerID, "reported", bc.TotalMemoryGB, "clamped", v)
-		bc.TotalMemoryGB = v
-	}
-	if v, changed := clampNonNeg(bc.GPUMemoryActiveGB, maxMemoryGBFloat); changed {
-		logger.Warn("provider gpu_memory_active_gb out of range, clamping",
-			"provider_id", providerID, "reported", bc.GPUMemoryActiveGB, "clamped", v)
-		bc.GPUMemoryActiveGB = v
-	}
-	if v, changed := clampNonNeg(bc.GPUMemoryPeakGB, maxMemoryGBFloat); changed {
-		bc.GPUMemoryPeakGB = v
-	}
-	if v, changed := clampNonNeg(bc.GPUMemoryCacheGB, maxMemoryGBFloat); changed {
-		bc.GPUMemoryCacheGB = v
-	}
-	for i := range bc.Slots {
-		s := &bc.Slots[i]
-		if s.MaxTokensPotential < 0 || s.MaxTokensPotential > maxTokensPotential {
-			logger.Warn("provider slot max_tokens_potential out of range, clamping",
-				"provider_id", providerID, "model", s.Model, "reported", s.MaxTokensPotential)
-			if s.MaxTokensPotential < 0 {
-				s.MaxTokensPotential = 0
-			} else {
-				s.MaxTokensPotential = maxTokensPotential
-			}
-		}
-		if s.NumRunning < 0 {
-			s.NumRunning = 0
-		}
-		if s.NumWaiting < 0 {
-			s.NumWaiting = 0
-		}
-		if s.MaxConcurrency < 0 || s.MaxConcurrency > maxReportedMaxConcurrency {
-			logger.Warn("provider slot max_concurrency out of range, clamping",
-				"provider_id", providerID, "model", s.Model, "reported", s.MaxConcurrency)
-			if s.MaxConcurrency < 0 {
-				s.MaxConcurrency = 0
-			} else {
-				s.MaxConcurrency = maxReportedMaxConcurrency
-			}
-		}
-		if v, changed := clampNonNeg(s.ObservedDecodeTPS, maxDecodeTPS); changed {
-			logger.Warn("provider slot observed_decode_tps out of range, clamping",
-				"provider_id", providerID, "model", s.Model, "reported", s.ObservedDecodeTPS, "clamped", v)
-			s.ObservedDecodeTPS = v
-		}
-		if s.ActiveTokenBudgetUsed < 0 || s.ActiveTokenBudgetUsed > maxTokenBudgetCap {
-			if s.ActiveTokenBudgetUsed < 0 {
-				s.ActiveTokenBudgetUsed = 0
-			} else {
-				s.ActiveTokenBudgetUsed = maxTokenBudgetCap
-			}
-		}
-		if s.ActiveTokenBudgetMax < 0 || s.ActiveTokenBudgetMax > maxTokenBudgetCap {
-			if s.ActiveTokenBudgetMax < 0 {
-				s.ActiveTokenBudgetMax = 0
-			} else {
-				s.ActiveTokenBudgetMax = maxTokenBudgetCap
-			}
-		}
-		if s.QueuedTokenBudget < 0 || s.QueuedTokenBudget > maxTokenBudgetCap {
-			if s.QueuedTokenBudget < 0 {
-				s.QueuedTokenBudget = 0
-			} else {
-				s.QueuedTokenBudget = maxTokenBudgetCap
-			}
-		}
-	}
-}
-
-// Register adds a new provider to the registry, returning its assigned ID.
-// Provider-reported model inventory is preserved even when the current catalog
-// denies every model; catalog checks are applied dynamically during routing so
-// providers that connect before a model is promoted become routable immediately
-// after the catalog is updated.
-func (r *Registry) Register(id string, conn *websocket.Conn, msg *protocol.RegisterMessage) *Provider {
-	// Clamp provider-reported performance stats used in routing score.
-	// Refuse to trust unbounded values — a malicious provider reporting
-	// DecodeTPS=1e9 would otherwise starve all other providers.
-	if v, changed := clampNonNeg(msg.DecodeTPS, maxDecodeTPS); changed {
-		r.logger.Warn("provider decode_tps out of range, clamping",
-			"provider_id", id, "reported", msg.DecodeTPS, "clamped", v)
-		msg.DecodeTPS = v
-	}
-	if v, changed := clampNonNeg(msg.PrefillTPS, maxPrefillTPS); changed {
-		r.logger.Warn("provider prefill_tps out of range, clamping",
-			"provider_id", id, "reported", msg.PrefillTPS, "clamped", v)
-		msg.PrefillTPS = v
-	}
-	if v, changed := clampNonNeg(msg.Hardware.MemoryBandwidthGBs, maxMemoryBandwidthGBs); changed {
-		r.logger.Warn("provider memory_bandwidth_gbs out of range, clamping",
-			"provider_id", id, "reported", msg.Hardware.MemoryBandwidthGBs, "clamped", v)
-		msg.Hardware.MemoryBandwidthGBs = v
-	}
-	if msg.Hardware.MemoryGB < 0 || msg.Hardware.MemoryGB > maxMemoryGB {
-		r.logger.Warn("provider memory_gb out of range, clamping",
-			"provider_id", id, "reported", msg.Hardware.MemoryGB)
-		if msg.Hardware.MemoryGB < 0 {
-			msg.Hardware.MemoryGB = 0
-		} else {
-			msg.Hardware.MemoryGB = maxMemoryGB
-		}
-	}
-
-	models := msg.Models
-
-	// Validate X25519 public key if provided.
-	// Reject invalid keys at registration rather than failing at encryption time.
-	pubKey := msg.PublicKey
-	if pubKey != "" {
-		decoded, err := base64.StdEncoding.DecodeString(pubKey)
-		if err != nil || len(decoded) != 32 {
-			r.logger.Warn("provider public key invalid, clearing",
-				"provider_id", id,
-				"error", "must be 32-byte base64-encoded X25519 key",
-			)
-			pubKey = "" // clear so provider can register but won't receive encrypted requests
-		}
-	}
-
-	p := &Provider{
-		ID:                      id,
-		Hardware:                msg.Hardware,
-		Models:                  models,
-		Backend:                 msg.Backend,
-		PublicKey:               pubKey,
-		EncryptedResponseChunks: msg.EncryptedResponseChunks,
-		PrefillTPS:              msg.PrefillTPS,
-		DecodeTPS:               msg.DecodeTPS,
-		TrustLevel:              TrustNone,
-		RuntimeVerified:         true,  // default to verified; API layer sets false when manifest check fails
-		RuntimeManifestChecked:  true,  // default to true; API layer sets false when no manifest is configured
-		ChallengeVerifiedSIP:    false, // starts false; set true by attestation challenge handler after SIP check
-		PrivacyCapabilities:     msg.PrivacyCapabilities,
-		TemplateHashes:          CloneStringMap(msg.TemplateHashes),
-		Status:                  StatusOnline,
-		Conn:                    conn,
-		LastHeartbeat:           time.Now(),
-		Reputation:              NewReputation(),
-		pendingReqs:             make(map[string]*PendingRequest),
-	}
-
-	r.mu.Lock()
-	r.providers[id] = p
-	r.onlineCount.Add(1)
-	for _, m := range models {
-		r.modelProviderInc(m.ID)
-	}
-	r.mu.Unlock()
-
-	r.logger.Info("provider registered",
-		"provider_id", id,
-		"chip", msg.Hardware.ChipName,
-		"memory_gb", msg.Hardware.MemoryGB,
-		"models", len(msg.Models),
-		"backend", msg.Backend,
-		"prefill_tps", msg.PrefillTPS,
-		"decode_tps", msg.DecodeTPS,
-	)
-
-	// Persist provider record to store (async).
-	r.persistProviderNow(p)
-
-	return p
-}
-
-func CloneStringMap(in map[string]string) map[string]string {
-	if len(in) == 0 {
-		return nil
-	}
-	out := make(map[string]string, len(in))
-	for k, v := range in {
-		out[k] = v
-	}
-	return out
-}
-
-// DisconnectDuplicatesBySerial disconnects all providers that share the same
-// serial number as the given provider, except the given provider itself.
-// This prevents multiple WebSocket connections from the same physical machine
-// from competing for the same vllm-mlx backend on localhost.
-func (r *Registry) DisconnectDuplicatesBySerial(keepID string, serial string) {
-	if serial == "" {
-		return
-	}
-
-	var toEvict []string
-
-	r.mu.RLock()
-	for id, p := range r.providers {
-		if id == keepID {
-			continue
-		}
-		if p.AttestationResult != nil && p.AttestationResult.SerialNumber == serial {
-			toEvict = append(toEvict, id)
-		}
-	}
-	r.mu.RUnlock()
-
-	for _, id := range toEvict {
-		r.mu.RLock()
-		p := r.providers[id]
-		r.mu.RUnlock()
-
-		r.logger.Warn("evicting duplicate provider from same device",
-			"evicted_id", id,
-			"kept_id", keepID,
-			"serial", serial,
-		)
-		r.Disconnect(id)
-
-		if p != nil && p.Conn != nil {
-			p.Conn.Close(websocket.StatusNormalClosure, "replaced by new connection from same device")
-		}
-	}
-}
-
-// Heartbeat updates the provider's status and stats.
-func (r *Registry) Heartbeat(id string, msg *protocol.HeartbeatMessage) {
-	r.mu.RLock()
-	p, ok := r.providers[id]
-	r.mu.RUnlock()
-	if !ok {
-		r.logger.Warn("heartbeat from unknown provider", "provider_id", id)
-		return
-	}
-
-	// Clamp heartbeat-reported capacity and metrics so a malicious provider
-	// can't skew routing by reporting absurd values (e.g. TotalMemoryGB=1e9
-	// would drive gpuUtil to 0 and sidestep health penalties).
-	clampBackendCapacity(r.logger, id, msg.BackendCapacity)
-	if v, changed := clampNonNeg(msg.SystemMetrics.MemoryPressure, 1.0); changed {
-		msg.SystemMetrics.MemoryPressure = v
-	}
-	if v, changed := clampNonNeg(msg.SystemMetrics.CPUUsage, 1.0); changed {
-		msg.SystemMetrics.CPUUsage = v
-	}
-
-	p.mu.Lock()
-	p.LastHeartbeat = time.Now()
-	p.Stats.RequestsServed += cumulativeDelta(p.lastSessionStats.RequestsServed, msg.Stats.RequestsServed)
-	p.Stats.TokensGenerated += cumulativeDelta(p.lastSessionStats.TokensGenerated, msg.Stats.TokensGenerated)
-	p.lastSessionStats = msg.Stats
-	p.SystemMetrics = msg.SystemMetrics
-	// Update backend capacity from heartbeat. A nil report clears prior live
-	// capacity so stale slot state cannot keep influencing routing.
-	p.BackendCapacity = msg.BackendCapacity
-	if p.BackendCapacity != nil {
-		chipFamily := p.Hardware.ChipFamily
-		for _, slot := range p.BackendCapacity.Slots {
-			if slot.ObservedDecodeTPS > 0 {
-				r.tpsRegistry.Record(slot.Model, chipFamily, slot.ObservedDecodeTPS)
-			}
-		}
-	}
-	// Update warm models from heartbeat. Always overwrite -- an empty list
-	// means the provider has no models loaded, and stale entries must be
-	// cleared to prevent TriggerModelSwaps from suppressing needed swaps.
-	p.WarmModels = msg.WarmModels
-	if msg.ActiveModel != nil {
-		p.CurrentModel = *msg.ActiveModel
-	} else {
-		// nil active_model means no model is loaded — clear stale state
-		// so attestation challenges don't compare against an unloaded model.
-		p.CurrentModel = ""
-	}
-	// Only update status from heartbeat if provider is not actively serving
-	// (serving status is managed by request lifecycle). Crucially, an
-	// untrusted provider must NOT transition back to StatusOnline here —
-	// that would cause an onlineCount double-decrement when Disconnect
-	// later sees StatusOnline and decrements a second time.
-	if p.Status == StatusUntrusted {
-		// no status transitions allowed
-	} else if p.Status != StatusServing || msg.Status == "idle" {
-		switch msg.Status {
-		case "idle":
-			p.Status = StatusOnline
-		case "serving":
-			p.Status = StatusServing
-		}
-	}
-	p.mu.Unlock()
-
-	r.PersistProviderThrottled(p)
-
-	// Heartbeats can make a recovered slot routable again (for example after a
-	// crash auto-restart). Drain matching queues using the canonical scheduler
-	// rather than the legacy direct queue assignment path.
-	r.drainQueuedRequestsForModels(providerModelIDs(p))
-
-	// If queue drain didn't satisfy all pending requests (no warm provider),
-	// check if a cold provider should swap models to serve queued demand.
-	r.TriggerModelSwaps()
-}
-
-// SendLoadModel instructs a provider to eagerly load a model so it becomes
-// warm for incoming requests. The provider will autonomously evict idle
-// models to make room. This is a fire-and-forget call — the coordinator
-// does not block waiting for the load to complete. The provider replies
-// asynchronously with a load_model_status message.
-func (r *Registry) SendLoadModel(providerID, modelID string) error {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return fmt.Errorf("provider %q not found", providerID)
-	}
-
-	msg := protocol.LoadModelMessage{
-		Type:    protocol.TypeLoadModel,
-		ModelID: modelID,
-	}
-	data, err := json.Marshal(msg)
-	if err != nil {
-		return fmt.Errorf("failed to marshal load_model message: %w", err)
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	p.mu.Lock()
-	conn := p.Conn
-	p.mu.Unlock()
-
-	if conn == nil {
-		return fmt.Errorf("provider %q has no active connection", providerID)
-	}
-
-	if err := conn.Write(ctx, websocket.MessageText, data); err != nil {
-		return fmt.Errorf("failed to send load_model to provider %q: %w", providerID, err)
-	}
-
-	r.logger.Info("sent load_model to provider",
-		"provider_id", providerID,
-		"model_id", modelID,
-	)
-	return nil
-}
-
-// TriggerModelSwaps checks for queued requests that have no warm provider
-// and sends load_model to cold providers that have the model available on
-// disk. This enables demand-driven model swapping: when requests queue for
-// a model that no provider has warm, the coordinator proactively triggers
-// a swap on an idle provider.
-//
-// Called after heartbeat processing and queue drain to catch demand that
-// can't be satisfied by warm providers alone.
-func (r *Registry) TriggerModelSwaps() {
-	if r.queue == nil {
-		return
-	}
-
-	queuedModels := r.queue.QueuedModels()
-	if len(queuedModels) == 0 {
-		return
-	}
-
-	now := time.Now()
-	r.expirePendingModelLoads(now)
-
-	actions := r.planModelLoadActions(queuedModels, now)
-	actions = r.reservePendingModelLoads(actions, now)
-	r.sendModelLoadActions(actions)
-}
-
-func (r *Registry) expirePendingModelLoads(now time.Time) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	for key, sentAt := range r.pendingModelLoads {
-		if now.Sub(sentAt) > pendingModelLoadTTL {
-			delete(r.pendingModelLoads, key)
-		}
-	}
-}
-
-func (r *Registry) planModelLoadActions(queuedModels []string, now time.Time) []modelLoadAction {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	selectedProviders := make(map[string]struct{})
-	actions := make([]modelLoadAction, 0, len(queuedModels))
-	for _, model := range queuedModels {
-		if r.hasWarmProviderLocked(model, now) {
-			continue
-		}
-
-		providerID := r.bestModelLoadProviderLocked(model, now, selectedProviders)
-		if providerID == "" {
-			continue
-		}
-		selectedProviders[providerID] = struct{}{}
-		actions = append(actions, modelLoadAction{providerID: providerID, modelID: model})
-	}
-	return actions
-}
-
-// hasWarmProviderLocked reports whether a connected provider already has the
-// model warm. Caller must hold r.mu (read or write).
-func (r *Registry) hasWarmProviderLocked(model string, now time.Time) bool {
-	for _, p := range r.providers {
-		p.mu.Lock()
-		warm := r.providerHasWarmModelLocked(p, model, now)
-		p.mu.Unlock()
-		if warm {
-			return true
-		}
-	}
-	return false
-}
-
-// providerHasWarmModelLocked checks whether the provider has the model warm
-// AND passes the same routing safety gates used by the scheduler. A provider
-// with stale attestation or failed privacy checks should not suppress swap
-// planning. Caller must hold p.mu. Caller must hold r.mu (read or write).
-func (r *Registry) providerHasWarmModelLocked(p *Provider, model string, now time.Time) bool {
-	if p.Status == StatusOffline || p.Status == StatusUntrusted {
-		return false
-	}
-	if trustRank(p.TrustLevel) < trustRank(r.MinTrustLevel) {
-		return false
-	}
-	if !p.RuntimeVerified {
-		return false
-	}
-	if !providerSupportsPrivateTextLocked(p) {
-		return false
-	}
-	if p.LastChallengeVerified.IsZero() || now.Sub(p.LastChallengeVerified) > challengeFreshnessMaxAge {
-		return false
-	}
-	if !r.providerServesCatalogModelLocked(p, model) {
-		return false
-	}
-	if p.BackendCapacity != nil {
-		for _, slot := range p.BackendCapacity.Slots {
-			if slot.Model == model {
-				// BackendCapacity is authoritative when present.
-				// Only "running" and "idle" mean the model is warm.
-				return slot.State == "running" || slot.State == "idle"
-			}
-		}
-		// Model has no slot in BackendCapacity -- it's not loaded.
-		return false
-	}
-	// Legacy provider without BackendCapacity: fall back to WarmModels.
-	for _, warmModel := range p.WarmModels {
-		if warmModel == model {
-			return true
-		}
-	}
-	return false
-}
-
-// bestModelLoadProviderLocked selects the eligible provider with the fewest
-// pending requests. Caller must hold r.mu (read or write).
-func (r *Registry) bestModelLoadProviderLocked(model string, now time.Time, selectedProviders map[string]struct{}) string {
-	bestProviderID := ""
-	for id, p := range r.providers {
-		if _, selected := selectedProviders[id]; selected {
-			continue
-		}
-		// Skip providers that have any pending model load -- sending a
-		// second load_model while the first is in progress can cause
-		// swap oscillation on single-slot providers.
-		if r.providerHasPendingLoad(id) {
-			continue
-		}
-
-		pendingCount, ok := r.modelLoadCandidatePendingLocked(p, model, now)
-		if !ok {
-			continue
-		}
-		// Only consider idle providers (no in-flight requests). Sending
-		// load_model to a provider that is actively serving another model
-		// will fail because the active slot cannot be evicted.
-		if pendingCount == 0 {
-			bestProviderID = id
-			break
-		}
-	}
-	return bestProviderID
-}
-
-// modelLoadCandidatePendingLocked applies the same routing safety gates used by
-// the scheduler, then returns the provider's current pending request count.
-// Caller must hold r.mu (read or write).
-func (r *Registry) modelLoadCandidatePendingLocked(p *Provider, model string, now time.Time) (int, bool) {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-
-	if p.Status == StatusOffline || p.Status == StatusUntrusted {
-		return 0, false
-	}
-	if trustRank(p.TrustLevel) < trustRank(r.MinTrustLevel) {
-		return 0, false
-	}
-	if !p.RuntimeVerified {
-		return 0, false
-	}
-	if !providerSupportsPrivateTextLocked(p) {
-		return 0, false
-	}
-	if p.LastChallengeVerified.IsZero() || now.Sub(p.LastChallengeVerified) > challengeFreshnessMaxAge {
-		return 0, false
-	}
-	if !r.providerServesCatalogModelLocked(p, model) {
-		return 0, false
-	}
-
-	// Memory gate: reject providers that cannot physically load the model.
-	// The provider applies a 2x multiplier on model weight size when deciding
-	// whether to load (ensureModelLoaded uses estimatedMemoryGb * 2.0 for
-	// headroom). Use the catalog SizeGB * 2.5 as the coordinator-side gate
-	// (slightly less conservative since the provider will reject anyway if
-	// it truly doesn't fit). This prevents the coordinator from sending
-	// load_model commands to machines that will always OOM-reject them.
-	if entry, ok := r.modelCatalog[model]; ok && entry.SizeGB > 0 {
-		requiredGB := entry.SizeGB * 2.5
-		providerMemGB := float64(p.Hardware.MemoryGB)
-		if providerMemGB > 0 && requiredGB > providerMemGB {
-			return 0, false
-		}
-	}
-
-	return p.pendingCount(), true
-}
-
-func (r *Registry) reservePendingModelLoads(actions []modelLoadAction, now time.Time) []modelLoadAction {
-	if len(actions) == 0 {
-		return nil
-	}
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	if r.pendingModelLoads == nil {
-		r.pendingModelLoads = make(map[string]time.Time)
-	}
-
-	reserved := actions[:0]
-	for _, action := range actions {
-		// Check per-provider (not just per-key) to prevent concurrent
-		// heartbeat goroutines from reserving the same idle provider
-		// for different models.
-		if r.providerHasPendingLoad(action.providerID) {
-			continue
-		}
-		r.pendingModelLoads[modelLoadKey(action.providerID, action.modelID)] = now
-		reserved = append(reserved, action)
-	}
-	return reserved
-}
-
-func (r *Registry) sendModelLoadActions(actions []modelLoadAction) {
-	for _, action := range actions {
-		if err := r.SendLoadModel(action.providerID, action.modelID); err != nil {
-			r.logger.Warn("failed to trigger model swap",
-				"provider_id", action.providerID,
-				"model_id", action.modelID,
-				"error", err,
-			)
-			r.ClearPendingModelLoad(action.providerID, action.modelID)
-		}
-	}
-}
-
-func modelLoadKey(providerID, modelID string) string {
-	return providerID + ":" + modelID
-}
-
-// providerHasPendingLoad reports whether the provider has any pending
-// load_model command. Caller must hold r.mu (read or write).
-func (r *Registry) providerHasPendingLoad(providerID string) bool {
-	prefix := providerID + ":"
-	for key := range r.pendingModelLoads {
-		if len(key) > len(prefix) && key[:len(prefix)] == prefix {
-			return true
-		}
-	}
-	return false
-}
-
-// MarkModelWarm adds a model to the provider's WarmModels list if not already
-// present. Called when load_model_status:succeeded arrives before the next
-// heartbeat, so the scheduler sees the provider as warm during queue drain.
-func (r *Registry) MarkModelWarm(providerID, modelID string) {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return
-	}
-
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	for _, wm := range p.WarmModels {
-		if wm == modelID {
-			return // already warm
-		}
-	}
-	p.WarmModels = append(p.WarmModels, modelID)
-	p.CurrentModel = modelID
-
-	// Inject a synthetic "idle" slot into BackendCapacity so the scheduler
-	// sees the model as warm. Without this, the scheduler only checks
-	// BackendCapacity.Slots (not WarmModels) for Swift providers, and a
-	// stale snapshot without the new model's slot would treat it as cold
-	// until the next heartbeat arrives.
-	//
-	// We only add/update the new model's slot and leave existing slots
-	// untouched — the provider may have multiple model slots loaded
-	// simultaneously (maxModelSlots defaults to 3). The next heartbeat
-	// will provide the authoritative slot list.
-	if p.BackendCapacity != nil {
-		found := false
-		for i, slot := range p.BackendCapacity.Slots {
-			if slot.Model == modelID {
-				p.BackendCapacity.Slots[i].State = "idle"
-				found = true
-				break
-			}
-		}
-		if !found {
-			p.BackendCapacity.Slots = append(p.BackendCapacity.Slots, protocol.BackendSlotCapacity{
-				Model: modelID,
-				State: "idle",
-			})
-		}
-	}
-}
-
-// ClearPendingModelLoad removes a pending model load entry after a terminal
-// load_model_status response.
-func (r *Registry) ClearPendingModelLoad(providerID, modelID string) {
-	r.mu.Lock()
-	delete(r.pendingModelLoads, modelLoadKey(providerID, modelID))
-	r.mu.Unlock()
-}
-
-// RejectUnservableQueuedRequests checks whether any eligible provider can
-// serve the given model. If not, all queued requests for the model are
-// rejected immediately rather than waiting for the 120s queue timeout.
-// Called after a load_model failure to give consumers a fast error.
-func (r *Registry) RejectUnservableQueuedRequests(modelID string) {
-	if r.queue == nil {
-		return
-	}
-	if r.queue.QueueSize(modelID) == 0 {
-		return
-	}
-
-	// Check if any provider can still serve this model. Only reject when
-	// NO provider serves the model at all. If providers exist but are
-	// temporarily at capacity (capacityRejections > 0), the requests
-	// should wait — those providers may finish current work and become
-	// available.
-	candidates, capacityRejections := r.QuickCapacityCheck(modelID, 500, defaultRequestedMaxTokens)
-	if candidates > 0 || capacityRejections > 0 {
-		return
-	}
-
-	failed := r.queue.FailQueuedRequestsForModel(modelID)
-	if failed > 0 {
-		r.logger.Warn("rejected queued requests for unservable model",
-			"model_id", modelID,
-			"rejected", failed,
-		)
-	}
-}
-
-func cumulativeDelta(previous, current int64) int64 {
-	if current <= 0 {
-		return 0
-	}
-	if current >= previous {
-		return current - previous
-	}
-	// The provider process restarted and reset its in-memory counters.
-	return current
-}
-
-// Disconnect removes a provider from the registry and cleans up pending requests.
-func (r *Registry) Disconnect(id string) {
-	r.mu.Lock()
-	p, ok := r.providers[id]
-	if ok {
-		delete(r.providers, id)
-		// Clear any pending model load entries for this provider.
-		for key := range r.pendingModelLoads {
-			if len(key) > len(id)+1 && key[:len(id)+1] == id+":" {
-				delete(r.pendingModelLoads, key)
-			}
-		}
-		p.mu.Lock()
-		if p.Status != StatusUntrusted {
-			r.onlineCount.Add(-1)
-			for _, m := range p.Models {
-				r.modelProviderDec(m.ID)
-			}
-		}
-		p.mu.Unlock()
-	}
-	r.mu.Unlock()
-
-	if !ok {
-		return
-	}
-
-	// Close all pending request channels so consumers get errors.
-	p.mu.Lock()
-	for reqID, pr := range p.pendingReqs {
-		pr.ErrorCh <- protocol.InferenceErrorMessage{
-			Type:       protocol.TypeInferenceError,
-			RequestID:  reqID,
-			Error:      "provider disconnected",
-			StatusCode: 502,
-		}
-		close(pr.ChunkCh)
-		close(pr.CompleteCh)
-		close(pr.ErrorCh)
-	}
-	p.pendingReqs = make(map[string]*PendingRequest)
-	p.mu.Unlock()
-
-	r.logger.Info("provider disconnected", "provider_id", id)
-}
-
-// GetProvider returns a provider by ID, or nil if not found.
-func (r *Registry) GetProvider(id string) *Provider {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return r.providers[id]
-}
-
-// MarkUntrusted sets a provider's status to untrusted for a hard/security
-// reason (bad encrypted chunk, MDM/MDA failure, SIP disabled, binary or model
-// hash mismatch, serial impersonation, attestation failure). The deroute is
-// non-recoverable: the provider stays untrusted until it reconnects and
-// re-registers. This is the default for every direct deroute call site.
-func (r *Registry) MarkUntrusted(providerID string) {
-	r.markUntrusted(providerID, false)
-}
-
-// MarkUntrustedTransient sets a provider's status to untrusted for a *transient*
-// reason — MaxFailedChallenges consecutive missed-challenge timeouts (screen
-// sleep, network blip, momentary Secure Enclave inaccessibility). Unlike
-// MarkUntrusted, the provider remains eligible to self-recover: the challenge
-// loop keeps challenging it (see ChallengeShouldStop), and a subsequent fully
-// passing challenge (RecordChallengeSuccess) restores it to online.
-//
-// A passing challenge re-verifies signature, SIP, secure boot, binary hash,
-// model hash and runtime before RecordChallengeSuccess is reached, so using it
-// as the recovery trigger is safe.
-func (r *Registry) MarkUntrustedTransient(providerID string) {
-	r.markUntrusted(providerID, true)
-}
-
-// markUntrusted is the shared implementation. recoverable=true marks the untrust
-// as transiently recoverable; recoverable=false is a hard deroute.
-//
-// Transition rules:
-//   - not untrusted -> untrusted: decrement online/model counts, set status and
-//     the recoverable flag.
-//   - already untrusted + hard (recoverable=false): clear the flag. A hard
-//     reason always overrides/downgrades a previously-recoverable untrust.
-//   - already untrusted + transient (recoverable=true): leave the flag as-is, so
-//     a transient timeout can never *upgrade* a hard deroute to recoverable
-//     (matters for an in-flight challenge timeout that races a hard deroute).
-func (r *Registry) markUntrusted(providerID string, recoverable bool) {
-	r.mu.Lock()
-	p, ok := r.providers[providerID]
-	if !ok {
-		r.mu.Unlock()
-		return
-	}
-
-	p.mu.Lock()
-	if p.Status != StatusUntrusted {
-		r.onlineCount.Add(-1)
-		for _, m := range p.Models {
-			r.modelProviderDec(m.ID)
-		}
-		p.Status = StatusUntrusted
-		p.untrustedRecoverable = recoverable
-	} else if !recoverable {
-		p.untrustedRecoverable = false
-	}
-	failed := p.FailedChallenges // read under p.mu (the old code read this unlocked)
-	p.mu.Unlock()
-	r.mu.Unlock()
-
-	r.logger.Warn("provider marked as untrusted",
-		"provider_id", providerID,
-		"failed_challenges", failed,
-		"recoverable", recoverable,
-	)
-}
-
-// SetTrustLevel updates a provider's trust level (thread-safe).
-func (r *Registry) SetTrustLevel(providerID string, level TrustLevel) {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return
-	}
-	p.mu.Lock()
-	p.TrustLevel = level
-	p.mu.Unlock()
-
-	// Persist trust state.
-	r.persistProviderNow(p)
-}
-
-// RecordChallengeSuccess records a successful challenge-response verification.
-// A fully passing challenge re-verifies signature, SIP, secure boot, binary
-// hash, model hash and runtime (see verifyChallengeResponse) before this is
-// called, so it doubles as the recovery trigger for a *transiently* untrusted
-// provider.
-//
-// Returns true iff this call recovered a transiently-untrusted provider back to
-// online. The caller (verifyChallengeResponse) uses that to push a fresh
-// "online" trust_status so the provider clears its local untrusted state and
-// cancels the pending diagnostic auto-report it scheduled at deroute time.
-func (r *Registry) RecordChallengeSuccess(providerID string) bool {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return false
-	}
-
-	recovered := r.recoverIfTransientlyUntrusted(providerID, p)
-
-	p.mu.Lock()
-	p.LastChallengeVerified = time.Now()
-	p.FailedChallenges = 0
-	if !p.ChallengeVerifiedSIP {
-		p.ChallengeVerifiedSIP = true
-	}
-	p.Reputation.RecordChallengePass()
-	p.mu.Unlock()
-
-	// Persist challenge state and reputation.
-	r.persistProviderNow(p)
-	r.persistReputation(p)
-
-	if recovered {
-		r.logger.Info("provider recovered from transient deroute", "provider_id", providerID)
-	}
-
-	// A newly verified (or newly recovered) provider may unlock queued requests
-	// for any model it serves.
-	r.drainQueuedRequestsForModels(providerModelIDs(p))
-
-	return recovered
-}
-
-// recoverIfTransientlyUntrusted promotes a transiently-untrusted provider back
-// to online, mirroring markUntrusted's bookkeeping in reverse. Returns true iff
-// a transition occurred. It acquires r.mu (write) then p.mu — the same order as
-// markUntrusted/Register/Disconnect — so online/model counts stay consistent and
-// the path is deadlock-free.
-func (r *Registry) recoverIfTransientlyUntrusted(providerID string, p *Provider) bool {
-	// Cheap pre-check under p.mu only, so the common (non-recovery) success path
-	// never contends on the registry write lock.
-	p.mu.Lock()
-	eligible := p.Status == StatusUntrusted && p.untrustedRecoverable
-	p.mu.Unlock()
-	if !eligible {
-		return false
-	}
-
-	r.mu.Lock()
-	// Re-verify membership: RecordChallengeSuccess looked p up under RLock and
-	// released it, so Disconnect may have removed (or replaced) it since. A
-	// transiently-untrusted provider was already decremented out of the counts,
-	// and Disconnect does not decrement an untrusted provider, so incrementing a
-	// stale/removed pointer here would permanently corrupt onlineCount and
-	// modelProviders. Only recover the provider still registered under this ID.
-	if cur, ok := r.providers[providerID]; !ok || cur != p {
-		r.mu.Unlock()
-		return false
-	}
-	p.mu.Lock()
-	// Re-check under the write lock: a hard deroute may have intervened and
-	// cleared the recoverable flag between the pre-check and here.
-	if p.Status != StatusUntrusted || !p.untrustedRecoverable {
-		p.mu.Unlock()
-		r.mu.Unlock()
-		return false
-	}
-	r.onlineCount.Add(1)
-	for _, m := range p.Models {
-		r.modelProviderInc(m.ID)
-	}
-	p.Status = StatusOnline
-	p.untrustedRecoverable = false
-	p.mu.Unlock()
-	r.mu.Unlock()
-	return true
-}
-
-// RecordChallengeFailure records a failed challenge-response. Returns the
-// new consecutive failure count.
-//
-// When transientOnly is true (timeout — the provider didn't respond in time),
-// routing eligibility is preserved until MaxFailedChallenges consecutive
-// failures. A single transient timeout should not instantly deroute a provider
-// that was verified seconds ago.
-//
-// When transientOnly is false (security failure — wrong signature, SIP
-// disabled, binary hash mismatch, etc.), routing eligibility is cleared
-// immediately because the provider actively failed a security check.
-func (r *Registry) RecordChallengeFailure(providerID string, transientOnly bool) int {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return 0
-	}
-
-	p.mu.Lock()
-	p.FailedChallenges++
-	p.Reputation.RecordChallengeFail()
-	count := p.FailedChallenges
-
-	if !transientOnly {
-		// Security failure — clear routing eligibility immediately.
-		p.LastChallengeVerified = time.Time{}
-		p.ChallengeVerifiedSIP = false
-	} else if count >= MaxFailedChallenges {
-		// Transient failures only clear after hitting the threshold.
-		p.LastChallengeVerified = time.Time{}
-		p.ChallengeVerifiedSIP = false
-	}
-	p.mu.Unlock()
-
-	// Persist challenge state and reputation.
-	r.persistProviderNow(p)
-	r.persistReputation(p)
-
-	return count
-}
-
-// TrustMultiplier returns the trust multiplier for routing score calculation.
-func TrustMultiplier(t TrustLevel) float64 {
-	switch t {
-	case TrustHardware:
-		return 1.0
-	case TrustSelfSigned:
-		return 0.8
-	default:
-		return 0.5
-	}
-}
-
-// DefaultMaxConcurrent is the fallback concurrency limit for providers
-// that don't report backend capacity. Providers that report BackendCapacity
-// in heartbeats get a dynamic limit based on their total memory.
-const DefaultMaxConcurrent = 4
-
-// MaxConcurrentRequests is kept as an alias for backward compatibility
-// with tests and external code that reference the old constant name.
-const MaxConcurrentRequests = DefaultMaxConcurrent
-
-// ScoreProvider calculates a routing score for a provider.
-// Higher scores indicate better routing candidates.
-// Score = (1 - load) * decode_tps * trust_multiplier * reputation * warm_bonus * health_factor
-//
-// Uses dynamic max based on hardware when backend capacity is reported.
-func ScoreProvider(p *Provider, model string) float64 {
-	// Providers that have not passed runtime integrity verification score 0
-	// and should never be selected for routing.
-	p.mu.Lock()
-	runtimeVerified := p.RuntimeVerified
-	p.mu.Unlock()
-	if !runtimeVerified {
-		return 0
-	}
-
-	// Load: gradient from 0.0 (idle) to 1.0 (at max concurrency).
-	// Uses a positive provider-reported slot cap when present, otherwise the
-	// legacy provider-level dynamic max.
-	p.mu.Lock()
-	maxConc := p.maxConcurrencyForModelLocked(model)
-	pending := float64(p.pendingLoadForModelLocked(model))
-	p.mu.Unlock()
-	load := pending / float64(maxConc)
-	if load > 1.0 {
-		load = 1.0
-	}
-
-	// Snapshot mutable fields under lock. These are written by Heartbeat
-	// and SetTrustLevel from other goroutines.
-	p.mu.Lock()
-	decodeTPS := p.DecodeTPS
-	trustLevel := p.TrustLevel
-	warmModels := append([]string{}, p.WarmModels...)
-	currentModel := p.CurrentModel
-	sysMetrics := p.SystemMetrics
-	repScore := p.Reputation.Score()
-	backendCap := p.BackendCapacity
-	p.mu.Unlock()
-
-	// Base decode TPS — when not reported by the provider, estimate from
-	// hardware memory bandwidth using sqrt scaling. Linear bandwidth
-	// ratios (e.g. 546 vs 300 = 1.8x) create too much routing skew;
-	// sqrt dampens this to ~1.35x, giving faster hardware a mild
-	// preference while still distributing load across all providers.
-	if decodeTPS <= 0 {
-		bw := float64(p.Hardware.MemoryBandwidthGBs)
-		if bw > 0 {
-			decodeTPS = math.Sqrt(bw) // sqrt scaling: 546→23.4, 400→20, 300→17.3, 150→12.2
-		} else {
-			decodeTPS = 1.0
-		}
-	}
-
-	trustMul := TrustMultiplier(trustLevel)
-
-	// Warm model bonus: only applies when the provider is IDLE (no pending
-	// requests). This prevents a warm provider from monopolizing all traffic.
-	// Once a warm provider has any pending requests, cold providers compete
-	// on equal terms — a 20s parallel cold-start beats waiting in a serial
-	// queue behind a single warm provider.
-	warmBonus := 1.0
-	isIdle := pending == 0
-	if isIdle {
-		for _, wm := range warmModels {
-			if wm == model {
-				warmBonus = 1.5
-				break
-			}
-		}
-		if currentModel == model {
-			warmBonus = 1.5
-		}
-	}
-
-	// Cold-start / crash penalty: apply regardless of load. These represent
-	// providers whose backend is DOWN (not just cold in cache). Loading from
-	// idle_shutdown takes ~30s, crashed backends may not recover at all.
-	if backendCap != nil {
-		for _, slot := range backendCap.Slots {
-			if slot.Model == model {
-				switch slot.State {
-				case "idle_shutdown":
-					warmBonus = 0.1
-				case "crashed":
-					warmBonus = 0.05
-				}
-				break
-			}
-		}
-	}
-
-	// Health factor from live system metrics
-	m := sysMetrics
-
-	// Memory pressure: linear penalty. At 0.9 -> factor 0.1
-	memFactor := 1.0 - m.MemoryPressure
-	if memFactor < 0.1 {
-		memFactor = 0.1
-	}
-
-	// CPU usage: gentle penalty (max 50% reduction at full load)
-	cpuFactor := 1.0 - (m.CPUUsage * 0.5)
-
-	// Thermal: step penalties
-	thermalFactor := 1.0
-	switch m.ThermalState {
-	case "fair":
-		thermalFactor = 0.8
-	case "serious":
-		thermalFactor = 0.4
-	case "critical":
-		thermalFactor = 0.0
-	}
-
-	healthFactor := memFactor * cpuFactor * thermalFactor
-
-	// GPU memory pressure from backend capacity: penalize providers with
-	// high GPU utilization to prefer those with more headroom.
-	if backendCap != nil && backendCap.GPUMemoryActiveGB > 0 {
-		totalMem := backendCap.TotalMemoryGB
-		if totalMem <= 0 {
-			totalMem = float64(p.Hardware.MemoryGB)
-		}
-		if totalMem > 0 {
-			gpuUtil := backendCap.GPUMemoryActiveGB / totalMem
-			gpuFactor := 1.0 - (gpuUtil * 0.5) // max 50% penalty at full GPU
-			if gpuFactor < 0.1 {
-				gpuFactor = 0.1
-			}
-			healthFactor *= gpuFactor
-		}
-	}
-
-	return (1.0 - load) * decodeTPS * trustMul * repScore * warmBonus * healthFactor
-}
-
-// FindProvider selects an available provider for the given model using
-// intelligent scoring based on benchmark data, trust level, reputation,
-// warm model cache, and backend capacity. Picks the highest-scoring
-// provider that has concurrency headroom (dynamic limit based on hardware).
-// Optional excludeIDs are provider IDs to skip (e.g. providers that
-// already failed for this request during retry).
-func (r *Registry) FindProvider(model string, excludeIDs ...string) *Provider {
-	return r.FindProviderWithTrust(model, "", excludeIDs...)
-}
-
-// FindProviderWithTrust selects a provider with an optional per-request
-// minimum trust level. If minTrust is empty, the registry's default
-// MinTrustLevel is used. Consumers can request a specific trust level
-// (e.g. hardware) to filter providers. Optional excludeIDs are provider
-// IDs to skip during selection.
-func (r *Registry) FindProviderWithTrust(model string, minTrust TrustLevel, excludeIDs ...string) *Provider {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	// Build a set of excluded provider IDs for O(1) lookup.
-	excludeSet := make(map[string]struct{}, len(excludeIDs))
-	for _, id := range excludeIDs {
-		excludeSet[id] = struct{}{}
-	}
-
-	// Determine effective minimum: max of registry default and per-request
-	effectiveMin := r.MinTrustLevel
-	if minTrust != "" && trustRank(minTrust) > trustRank(effectiveMin) {
-		effectiveMin = minTrust
-	}
-
-	// Challenge staleness threshold: providers must have passed a
-	// challenge within the last interval + grace period. The challenge
-	// interval is 5 minutes, so we allow up to 6 minutes (interval +
-	// 1-minute grace) to avoid a gap where providers are unroutable
-	// between challenge cycles.
-	challengeMaxAge := 6 * time.Minute
-	now := time.Now()
-
-	var candidates []*Provider
-	for _, p := range r.providers {
-		// Skip explicitly excluded providers (failed on previous retry attempts).
-		if _, excluded := excludeSet[p.ID]; excluded {
-			continue
-		}
-
-		p.mu.Lock()
-		status := p.Status
-		trust := p.TrustLevel
-		lastChallenge := p.LastChallengeVerified
-		runtimeVerified := p.RuntimeVerified
-		privateReady := providerSupportsPrivateTextLocked(p)
-		p.mu.Unlock()
-
-		if status == StatusOffline || status == StatusUntrusted {
-			continue
-		}
-		if trustRank(trust) < trustRank(effectiveMin) {
-			continue
-		}
-		if !runtimeVerified || !privateReady {
-			continue
-		}
-		if lastChallenge.IsZero() || now.Sub(lastChallenge) > challengeMaxAge {
-			continue
-		}
-		p.mu.Lock()
-		hasHeadroom := p.hasConcurrencyHeadroomForModelLocked(model)
-		p.mu.Unlock()
-		if !hasHeadroom {
-			continue
-		}
-		if r.providerServesCatalogModelLocked(p, model) {
-			candidates = append(candidates, p)
-		}
-	}
-
-	if len(candidates) == 0 {
-		return nil
-	}
-
-	// Score all candidates and pick the highest.
-	bestIdx := 0
-	bestScore := ScoreProvider(candidates[0], model)
-	for i := 1; i < len(candidates); i++ {
-		s := ScoreProvider(candidates[i], model)
-		if s > bestScore {
-			bestScore = s
-			bestIdx = i
-		}
-	}
-
-	// When multiple candidates tie for the best score (common when all
-	// providers have the same hardware/TPS and load), randomly pick among
-	// them to distribute load instead of always picking the first one.
-	var ties []*Provider
-	for _, c := range candidates {
-		if ScoreProvider(c, model) >= bestScore-0.001 {
-			ties = append(ties, c)
-		}
-	}
-	var selected *Provider
-	if len(ties) > 1 {
-		selected = ties[rand.Intn(len(ties))]
-	} else {
-		selected = candidates[bestIdx]
-	}
-
-	selected.mu.Lock()
-	selected.Status = StatusServing
-	selected.mu.Unlock()
-
-	return selected
-}
-
-// SetProviderIdle updates a provider's status after a request completes.
-// If pending count reaches zero, status goes back to online. If there are
-// queued requests and the provider has concurrency headroom, the next
-// queued request is assigned immediately.
-func (r *Registry) SetProviderIdle(id string) {
-	r.mu.RLock()
-	p, ok := r.providers[id]
-	r.mu.RUnlock()
-	if !ok {
-		return
-	}
-
-	p.mu.Lock()
-	if p.pendingCount() == 0 && p.Status != StatusUntrusted && p.Status != StatusOffline {
-		p.Status = StatusOnline
-	}
-	p.mu.Unlock()
-
-	// Use all newly available capacity, not just a single queued request.
-	r.drainQueuedRequestsForModels(providerModelIDs(p))
-}
-
-// AttestationSummary provides aggregate attestation status for a model's providers.
-type AttestationSummary struct {
-	SecureEnclave bool `json:"secure_enclave"`
-	SIPEnabled    bool `json:"sip_enabled"`
-	SecureBoot    bool `json:"secure_boot"`
-}
-
-// AggregateModel is a deduplicated model entry for the /v1/models endpoint.
-type AggregateModel struct {
-	ID                string              `json:"id"`
-	ModelType         string              `json:"model_type"`
-	Quantization      string              `json:"quantization"`
-	Providers         int                 `json:"providers"`          // number of providers offering this model
-	AttestedProviders int                 `json:"attested_providers"` // number of attested providers
-	TrustLevel        TrustLevel          `json:"trust_level"`        // highest trust level among providers
-	Attestation       *AttestationSummary `json:"attestation,omitempty"`
-}
-
-// ListModels returns deduplicated models from all online providers.
-func (r *Registry) ListModels() []AggregateModel {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	type modelAgg struct {
-		modelType     string
-		quantization  string
-		count         int
-		attestedCount int
-		highestTrust  TrustLevel
-		secureEnclave bool
-		sipEnabled    bool
-		secureBoot    bool
-	}
-
-	// Aggregate by model ID only — consumers request by ID, so providers
-	// offering the same model ID should be counted together regardless of
-	// minor metadata differences.
-	agg := make(map[string]*modelAgg)
-	for _, p := range r.providers {
-		p.mu.Lock()
-		status := p.Status
-		trust := p.TrustLevel
-		attested := p.Attested
-		attestResult := p.AttestationResult
-		privateReady := providerSupportsPrivateTextLocked(p)
-		p.mu.Unlock()
-
-		if status == StatusOffline || status == StatusUntrusted {
-			continue
-		}
-		if !r.trustMeetsMinimum(trust) || !privateReady {
-			continue
-		}
-		for _, m := range p.Models {
-			if !r.modelAllowedByCatalogLocked(m) {
-				continue
-			}
-			k := m.ID
-			a, ok := agg[k]
-			if !ok {
-				a = &modelAgg{
-					modelType:    m.ModelType,
-					quantization: m.Quantization,
-					highestTrust: TrustNone,
-				}
-				agg[k] = a
-			}
-			a.count++
-
-			// Update highest trust level
-			if trustRank(trust) > trustRank(a.highestTrust) {
-				a.highestTrust = trust
-			}
-
-			if attested && attestResult != nil {
-				a.attestedCount++
-				a.secureEnclave = a.secureEnclave || attestResult.SecureEnclaveAvailable
-				a.sipEnabled = a.sipEnabled || attestResult.SIPEnabled
-				a.secureBoot = a.secureBoot || attestResult.SecureBootEnabled
-			}
-		}
-	}
-
-	models := make([]AggregateModel, 0, len(agg))
-	for k, a := range agg {
-		am := AggregateModel{
-			ID:                k,
-			ModelType:         a.modelType,
-			Quantization:      a.quantization,
-			Providers:         a.count,
-			AttestedProviders: a.attestedCount,
-			TrustLevel:        a.highestTrust,
-		}
-		if a.attestedCount > 0 {
-			am.Attestation = &AttestationSummary{
-				SecureEnclave: a.secureEnclave,
-				SIPEnabled:    a.sipEnabled,
-				SecureBoot:    a.secureBoot,
-			}
-		}
-		models = append(models, am)
-	}
-
-	return models
-}
-
-// ModelCountryCodes returns the sorted, de-duplicated ISO 3166-1 alpha-2
-// country codes of online providers serving the given model. Used to populate
-// the OpenRouter "datacenters" field. Only routing-eligible providers count —
-// the same gates as ListModels (online, meets the minimum trust level, and
-// private-text ready) — so a country whose providers can't actually serve the
-// model is not advertised. Providers without a known location are skipped.
-func (r *Registry) ModelCountryCodes(modelID string) []string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	seen := make(map[string]bool)
-	for _, p := range r.providers {
-		p.mu.Lock()
-		status := p.Status
-		trust := p.TrustLevel
-		privateReady := providerSupportsPrivateTextLocked(p)
-		var cc string
-		if p.Location != nil {
-			cc = strings.ToUpper(strings.TrimSpace(p.Location.CountryCode))
-		}
-		serves := false
-		if cc != "" {
-			for i := range p.Models {
-				if p.Models[i].ID == modelID {
-					serves = true
-					break
-				}
-			}
-		}
-		p.mu.Unlock()
-		if !serves {
-			continue
-		}
-		// Apply the same routing-eligibility gates as ListModels.
-		if status == StatusOffline || status == StatusUntrusted {
-			continue
-		}
-		if !r.trustMeetsMinimum(trust) || !privateReady {
-			continue
-		}
-		seen[cc] = true
-	}
-	if len(seen) == 0 {
-		return nil
-	}
-	out := make([]string, 0, len(seen))
-	for c := range seen {
-		out = append(out, c)
-	}
-	sort.Strings(out)
-	return out
-}
-
-// trustRank returns a numeric rank for trust levels (higher = more trusted).
-// Returns -1 for unknown/invalid trust levels.
-func trustRank(t TrustLevel) int {
-	switch t {
-	case TrustHardware:
-		return 2
-	case TrustSelfSigned:
-		return 1
-	case TrustNone:
-		return 0
-	default:
-		return -1
-	}
-}
-
-// RecordJobSuccess records a successful job completion for the provider's reputation.
-func (r *Registry) RecordJobSuccess(providerID string, responseTime time.Duration) {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return
-	}
-
-	p.mu.Lock()
-	p.Reputation.RecordJobSuccess(responseTime)
-	p.mu.Unlock()
-
-	// Persist reputation.
-	r.persistReputation(p)
-}
-
-// RecordJobFailure records a failed job for the provider's reputation.
-func (r *Registry) RecordJobFailure(providerID string) {
-	r.mu.RLock()
-	p, ok := r.providers[providerID]
-	r.mu.RUnlock()
-	if !ok {
-		return
-	}
-
-	p.mu.Lock()
-	p.Reputation.RecordJobFailure()
-	p.mu.Unlock()
-
-	// Persist reputation.
-	r.persistReputation(p)
-}
-
-// ProviderCount returns the number of registered providers.
-// modelProviderInc increments the provider count for a model. Must be called
-// with r.mu held.
-func (r *Registry) modelProviderInc(model string) {
-	r.modelProvidersMu.Lock()
-	c, ok := r.modelProviders[model]
-	if !ok {
-		c = &atomic.Int64{}
-		r.modelProviders[model] = c
-	}
-	r.modelProvidersMu.Unlock()
-	c.Add(1)
-}
-
-// modelProviderDec decrements the provider count for a model. Must be called
-// with r.mu held.
-func (r *Registry) modelProviderDec(model string) {
-	r.modelProvidersMu.Lock()
-	c, ok := r.modelProviders[model]
-	r.modelProvidersMu.Unlock()
-	if ok {
-		v := c.Add(-1)
-		if v <= 0 {
-			r.modelProvidersMu.Lock()
-			delete(r.modelProviders, model)
-			r.modelProvidersMu.Unlock()
-		}
-	}
-}
-
-// OnlineCount returns the number of online providers.
-func (r *Registry) OnlineCount() int64 {
-	return r.onlineCount.Load()
-}
-
-// ModelProviderSnapshot returns a snapshot of model_id -> provider count.
-func (r *Registry) ModelProviderSnapshot() map[string]int64 {
-	r.modelProvidersMu.Lock()
-	snap := make(map[string]int64, len(r.modelProviders))
-	for model, c := range r.modelProviders {
-		if v := c.Load(); v > 0 {
-			snap[model] = v
-		}
-	}
-	r.modelProvidersMu.Unlock()
-	return snap
-}
-
-func (r *Registry) ProviderCount() int {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return len(r.providers)
-}
-
-func (r *Registry) ProviderCountByVersion() map[string]int {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	counts := make(map[string]int)
-	for _, p := range r.providers {
-		p.mu.Lock()
-		online := p.Status != StatusOffline && p.Status != StatusUntrusted
-		p.mu.Unlock()
-		if !online {
-			continue
-		}
-		ver := p.Version
-		if ver == "" {
-			ver = "unknown"
-		}
-		counts[ver]++
-	}
-	return counts
-}
-
-// FleetSnapshot is the read-only summary used by metrics polling. We
-// don't lock individual providers — counts may be off-by-one under
-// heavy churn — that's acceptable for gauges.
-type FleetSnapshot struct {
-	Connected  int
-	Idle       int
-	QueueDepth int
-}
-
-// Snapshot returns aggregate counts for /metrics gauges. Cheap enough
-// to call every few seconds. Takes the registry's read lock for the
-// outer iteration AND each provider's mutex briefly to read Status and
-// pending count — those fields are written under p.mu elsewhere
-// (Heartbeat, AddPending, RemovePending), so reading them without
-// p.mu is a data race even if the gauge value is only advisory.
-func (r *Registry) Snapshot() FleetSnapshot {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	idle := 0
-	for _, p := range r.providers {
-		p.mu.Lock()
-		isIdle := p.Status == StatusOnline && len(p.pendingReqs) == 0
-		p.mu.Unlock()
-		if isIdle {
-			idle++
-		}
-	}
-	q := 0
-	if r.queue != nil {
-		q = r.queue.TotalSize()
-	}
-	return FleetSnapshot{
-		Connected:  len(r.providers),
-		Idle:       idle,
-		QueueDepth: q,
-	}
-}
-
-// ModelCapacity describes the live capacity for a single model.
-type ModelCapacity struct {
-	ModelID              string  `json:"id"`
-	Ready                bool    `json:"ready"`                  // at least one routable provider with headroom
-	CanAccept            bool    `json:"can_accept"`             // ready AND queue not full
-	RoutableProviders    int     `json:"routable_providers"`     // passed all gates
-	WarmProviders        int     `json:"warm_providers"`         // model loaded (slot state "running")
-	ColdProviders        int     `json:"cold_providers"`         // model available but not loaded
-	ActiveRequests       int     `json:"active_requests"`        // in-flight across fleet
-	QueuedRequests       int     `json:"queued_requests"`        // waiting in coordinator queue
-	QueueLimit           int     `json:"queue_limit"`            // max queue depth per model
-	AggregateTPS         float64 `json:"aggregate_tps"`          // sum of effective decode TPS
-	EstimatedTTFTMs      int64   `json:"estimated_ttft_ms"`      // best-case TTFT from lowest-cost warm provider
-	TokenBudgetRemaining int64   `json:"token_budget_remaining"` // aggregate free budget across providers
-	TokenBudgetTotal     int64   `json:"token_budget_total"`     // aggregate total budget
-}
-
-// providerCapSnap is a per-provider snapshot collected under the registry
-// lock, then aggregated into ModelCapacity outside the lock.
-type providerCapSnap struct {
-	model                 string
-	warm                  bool
-	hasHeadroom           bool // pending < maxConcurrency
-	effectiveTPS          float64
-	prefillTPS            float64
-	activeRequests        int // numRunning + numWaiting from backend slot, or pendingCount
-	backlogTokens         float64
-	activeTokenBudgetMax  int64
-	activeTokenBudgetUsed int64
-	queuedTokenBudget     int64
-}
-
-// ModelCapacitySnapshot returns a capacity snapshot for every model served
-// by at least one provider. Providers must pass the same routing gates as
-// snapshotProviderLocked (status, trust, runtime, privacy, challenge
-// freshness, concurrency headroom) to be counted as routable.
-func (r *Registry) ModelCapacitySnapshot() []ModelCapacity {
-	now := time.Now()
-
-	// Phase 1: collect per-provider snapshots under the lock.
-	var snaps []providerCapSnap
-
-	r.mu.RLock()
-	for _, p := range r.providers {
-		p.mu.Lock()
-
-		// Apply the same gates as snapshotProviderLocked.
-		if p.Status == StatusOffline || p.Status == StatusUntrusted {
-			p.mu.Unlock()
-			continue
-		}
-		if trustRank(p.TrustLevel) < trustRank(r.MinTrustLevel) {
-			p.mu.Unlock()
-			continue
-		}
-		if !p.RuntimeVerified {
-			p.mu.Unlock()
-			continue
-		}
-		if !providerSupportsPrivateTextLocked(p) {
-			p.mu.Unlock()
-			continue
-		}
-		if p.LastChallengeVerified.IsZero() || now.Sub(p.LastChallengeVerified) > challengeFreshnessMaxAge {
-			p.mu.Unlock()
-			continue
-		}
-
-		decodeTPS := resolvedDecodeTPS(p)
-		prefillTPS := resolvedPrefillTPS(p)
-
-		// Enumerate every model this provider serves.
-		for _, m := range p.Models {
-			if !r.modelAllowedByCatalogLocked(m) {
-				continue
-			}
-			hasHeadroom := p.hasConcurrencyHeadroomForModelLocked(m.ID)
-			// Count only pending requests for this specific model, not the
-			// total across all models. Using the total inflates
-			// activeRequests for multi-model providers.
-			modelPending := 0
-			for _, pr := range p.pendingReqs {
-				if pr.Model == m.ID {
-					modelPending++
-				}
-			}
-
-			snap := providerCapSnap{
-				model:          m.ID,
-				hasHeadroom:    hasHeadroom,
-				effectiveTPS:   decodeTPS,
-				prefillTPS:     prefillTPS,
-				activeRequests: modelPending,
-			}
-
-			// Check backend capacity for this model's slot.
-			if p.BackendCapacity != nil {
-				for _, slot := range p.BackendCapacity.Slots {
-					if slot.Model != m.ID {
-						continue
-					}
-					snap.warm = slot.State == "running"
-					slotActive := int(slot.NumRunning) + int(slot.NumWaiting)
-					if slotActive > snap.activeRequests {
-						snap.activeRequests = slotActive
-					}
-					if slot.ObservedDecodeTPS > 0 {
-						snap.effectiveTPS = slot.ObservedDecodeTPS
-					}
-					snap.activeTokenBudgetMax = slot.ActiveTokenBudgetMax
-					snap.activeTokenBudgetUsed = slot.ActiveTokenBudgetUsed
-					snap.queuedTokenBudget = slot.QueuedTokenBudget
-					snap.backlogTokens = float64(slot.MaxTokensPotential)
-					break
-				}
-			} else {
-				// Without backend capacity, warm if currently serving this model.
-				snap.warm = p.CurrentModel == m.ID
-			}
-
-			snaps = append(snaps, snap)
-		}
-		p.mu.Unlock()
-	}
-	r.mu.RUnlock()
-
-	// Phase 2: aggregate per-model outside the lock.
-	type modelAgg struct {
-		routable         int
-		warm             int
-		cold             int
-		activeRequests   int
-		aggregateTPS     float64
-		budgetRemaining  int64
-		budgetTotal      int64
-		bestWarmTTFTMs   int64 // -1 = not set
-		bestColdTTFTMs   int64 // -1 = not set
-		anyImmediateSlot bool  // at least one provider with headroom
-	}
-	agg := make(map[string]*modelAgg)
-	for _, s := range snaps {
-		a, ok := agg[s.model]
-		if !ok {
-			a = &modelAgg{bestWarmTTFTMs: -1, bestColdTTFTMs: -1}
-			agg[s.model] = a
-		}
-		if s.warm {
-			a.warm++
-		} else {
-			a.cold++
-		}
-		a.activeRequests += s.activeRequests
-		a.aggregateTPS += s.effectiveTPS
-		if s.activeTokenBudgetMax > 0 {
-			headroom := s.activeTokenBudgetMax - s.activeTokenBudgetUsed - s.queuedTokenBudget
-			if headroom < 0 {
-				headroom = 0
-			}
-			a.budgetRemaining += headroom
-			a.budgetTotal += s.activeTokenBudgetMax
-		}
-		// Routable providers require both concurrency headroom AND token-budget
-		// headroom. A provider with exhausted token budget should not make the
-		// model appear immediately ready.
-		hasBudgetHeadroom := s.activeTokenBudgetMax <= 0 ||
-			s.activeTokenBudgetUsed+s.queuedTokenBudget < s.activeTokenBudgetMax
-		if s.hasHeadroom && hasBudgetHeadroom {
-			a.routable++
-			a.anyImmediateSlot = true
-		}
-
-		// Estimate TTFT for this provider: prefill 500 tokens + backlog drain.
-		const defaultPromptTokens = 500
-		ttftMs := int64(0)
-		if s.prefillTPS > 0 {
-			ttftMs = int64(float64(defaultPromptTokens) / s.prefillTPS * 1000)
-		}
-		if s.effectiveTPS > 0 {
-			ttftMs += int64(s.backlogTokens / s.effectiveTPS * 1000)
-		}
-		if s.warm {
-			if a.bestWarmTTFTMs < 0 || ttftMs < a.bestWarmTTFTMs {
-				a.bestWarmTTFTMs = ttftMs
-			}
-		} else {
-			coldTTFT := ttftMs + 20_000 // 20s cold-start penalty
-			if a.bestColdTTFTMs < 0 || coldTTFT < a.bestColdTTFTMs {
-				a.bestColdTTFTMs = coldTTFT
-			}
-		}
-	}
-
-	// Phase 3: read queue sizes (separate lock, safe to call after releasing r.mu).
-	queueLimit := 0
-	if r.queue != nil {
-		queueLimit = r.queue.MaxSize()
-	}
-
-	result := make([]ModelCapacity, 0, len(agg))
-	for model, a := range agg {
-		queued := 0
-		if r.queue != nil {
-			queued = r.queue.QueueSize(model)
-		}
-		ready := a.routable > 0
-		canAccept := ready && (queued < queueLimit || a.anyImmediateSlot)
-
-		ttft := a.bestWarmTTFTMs
-		if ttft < 0 {
-			ttft = a.bestColdTTFTMs
-		}
-		if ttft < 0 {
-			ttft = 0
-		}
-
-		result = append(result, ModelCapacity{
-			ModelID:              model,
-			Ready:                ready,
-			CanAccept:            canAccept,
-			RoutableProviders:    a.routable,
-			WarmProviders:        a.warm,
-			ColdProviders:        a.cold,
-			ActiveRequests:       a.activeRequests,
-			QueuedRequests:       queued,
-			QueueLimit:           queueLimit,
-			AggregateTPS:         a.aggregateTPS,
-			EstimatedTTFTMs:      ttft,
-			TokenBudgetRemaining: a.budgetRemaining,
-			TokenBudgetTotal:     a.budgetTotal,
-		})
-	}
-	return result
-}
-
-// ForEachProvider iterates over all registered providers (read lock held).
-func (r *Registry) ForEachProvider(fn func(p *Provider)) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	for _, p := range r.providers {
-		fn(p)
-	}
-}
-
-// ProviderIDs returns the IDs of all registered providers.
-func (r *Registry) ProviderIDs() []string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	ids := make([]string, 0, len(r.providers))
-	for id := range r.providers {
-		ids = append(ids, id)
-	}
-	return ids
-}
-
-// StartEvictionLoop starts a background goroutine that removes providers
-// that haven't sent a heartbeat within the given timeout. It stops when
-// the context is cancelled.
-func (r *Registry) StartEvictionLoop(ctx context.Context, timeout time.Duration) {
-	ticker := time.NewTicker(timeout / 3)
-	saferun.Go(r.logger, "registry.evictionLoop", func() {
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-ticker.C:
-				r.evictStale(timeout)
-			}
-		}
-	})
-}
-
-func (r *Registry) evictStale(timeout time.Duration) {
-	r.mu.RLock()
-	var stale []string
-	now := time.Now()
-	for id, p := range r.providers {
-		if now.Sub(p.LastHeartbeat) > timeout {
-			stale = append(stale, id)
-		}
-	}
-	r.mu.RUnlock()
-
-	for _, id := range stale {
-		r.logger.Warn("evicting stale provider", "provider_id", id, "timeout", timeout)
-		r.Disconnect(id)
-	}
 }
