+++ coordinator/registry/snapshot.go
@@ -1,0 +1,78 @@
+package registry
+
+import (
+	"time"
+)
+
+func (r *Registry) snapshotProviderLocked(p *Provider, model string) (routingSnapshot, bool) {
+	now := time.Now()
+
+	p.mu.Lock()
+	defer p.mu.Unlock()
+
+	if !r.providerServesCatalogModelLocked(p, model) {
+		return routingSnapshot{}, false
+	}
+	if p.Status == StatusOffline || p.Status == StatusUntrusted {
+		return routingSnapshot{}, false
+	}
+	if trustRank(p.TrustLevel) < trustRank(r.MinTrustLevel) {
+		return routingSnapshot{}, false
+	}
+	if !p.RuntimeVerified {
+		return routingSnapshot{}, false
+	}
+	if !providerSupportsPrivateTextLocked(p) {
+		return routingSnapshot{}, false
+	}
+	if p.LastChallengeVerified.IsZero() || now.Sub(p.LastChallengeVerified) > challengeFreshnessMaxAge {
+		return routingSnapshot{}, false
+	}
+
+	snap := routingSnapshot{
+		provider:      p,
+		model:         model,
+		slotState:     "unknown",
+		totalPending:  p.pendingCount(),
+		systemMetrics: p.SystemMetrics,
+		decodeTPS:     resolvedDecodeTPS(p),
+		prefillTPS:    resolvedPrefillTPS(p),
+		totalMemoryGB: float64(p.Hardware.MemoryGB),
+		modelSizeGB:   r.catalogSizeGBLocked(model),
+	}
+
+	for _, pr := range p.pendingReqs {
+		if pr.Model != model {
+			continue
+		}
+		snap.pendingForModel++
+		snap.pendingMaxTokens += pendingTokenBudget(pr)
+	}
+	snap.hasHeadroom = p.hasConcurrencyHeadroomForModelLocked(model)
+
+	if p.BackendCapacity != nil {
+		snap.gpuMemoryActiveGB = p.BackendCapacity.GPUMemoryActiveGB
+		if p.BackendCapacity.TotalMemoryGB > 0 {
+			snap.totalMemoryGB = p.BackendCapacity.TotalMemoryGB
+		}
+		for _, slot := range p.BackendCapacity.Slots {
+			if slot.Model != model {
+				continue
+			}
+			snap.slotState = slot.State
+			snap.backendRunning = int(slot.NumRunning)
+			snap.backendWaiting = int(slot.NumWaiting)
+			snap.maxTokensPotential = slot.MaxTokensPotential
+			snap.observedDecodeTPS = slot.ObservedDecodeTPS
+			snap.activeTokenBudgetUsed = slot.ActiveTokenBudgetUsed
+			snap.activeTokenBudgetMax = slot.ActiveTokenBudgetMax
+			snap.queuedTokenBudget = slot.QueuedTokenBudget
+			break
+		}
+	}
+	snap.modelLoaded = snap.slotState == "running"
+	snap.availableOnDisk = !snap.modelLoaded
+	snap.fleetMedianTPS = r.tpsRegistry.Median(model, p.Hardware.ChipFamily)
+
+	return snap, true
+}
