+++ coordinator/registry/trust.go
@@ -1,0 +1,57 @@
+package registry
+
+// ProviderStatus represents the operational state of a provider.
+type ProviderStatus string
+
+const (
+	StatusOnline    ProviderStatus = "online"
+	StatusServing   ProviderStatus = "serving"
+	StatusOffline   ProviderStatus = "offline"
+	StatusUntrusted ProviderStatus = "untrusted"
+)
+
+// TrustLevel represents the attestation trust level of a provider.
+type TrustLevel string
+
+const (
+	TrustNone       TrustLevel = "none"        // No attestation provided
+	TrustSelfSigned TrustLevel = "self_signed" // Attestation signed by provider's own key
+	TrustHardware   TrustLevel = "hardware"    // MDM + MDA + SE key bound to Apple-verified hardware
+)
+
+const BackendMLXSwift = "mlx-swift"
+
+// MaxFailedChallenges is the number of consecutive challenge failures before
+// a provider is marked untrusted and fully derouted.
+const MaxFailedChallenges = 3
+
+func BackendUsesSwiftRuntime(backend string) bool {
+	return backend == BackendMLXSwift
+}
+
+// TrustMultiplier returns the trust multiplier for routing score calculation.
+func TrustMultiplier(t TrustLevel) float64 {
+	switch t {
+	case TrustHardware:
+		return 1.0
+	case TrustSelfSigned:
+		return 0.8
+	default:
+		return 0.5
+	}
+}
+
+// trustRank returns a numeric rank for trust levels (higher = more trusted).
+// Returns -1 for unknown/invalid trust levels.
+func trustRank(t TrustLevel) int {
+	switch t {
+	case TrustHardware:
+		return 2
+	case TrustSelfSigned:
+		return 1
+	case TrustNone:
+		return 0
+	default:
+		return -1
+	}
+}
